### 2022-07-08 #4559 @editorialbot
[PRE REVIEW]: Fastsubtrees: simple and efficient subtrees extractions in Python with applications to NCBI taxonomy
**Submitting author:** @ggonnella (<a href="http://orcid.org/0000-0003-3900-5397">Giorgio Gonnella</a>)
**Repository:** https://github.com/ggonnella/fastsubtrees
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0
