### 2022-07-08 4b3e6e3
add ntmirror requirement
