+++ requirements.txt
@@ -12,6 +12,6 @@ fastsubtrees~=0.2
 pytest~=7.1.1
 plotly~=5.5.0
 dash~=2.0.0
-ntmirror~=1.0.0
+ntmirror~=1.1.0
 setuptools~=47.1.0
 pandas~=1.1.4
