### 2022-07-08 b0f7b4b
remove requirements which are not neeeded for main library
