+++ requirements.txt
@@ -1,17 +1,10 @@
 docopt~=0.6.2
 loguru~=0.5.3
 tqdm~=4.62.3
-SQLAlchemy~=1.4.23
 schema~=0.7.4
 sh~=1.14.2
-nimporter~=2.0.0
-mysqlclient~=2.0.3
 PyYAML~=6.0
-numpy~=1.19.4
 fastsubtrees~=0.2
 pytest~=7.1.1
-plotly~=5.5.0
-dash~=2.0.0
 ntmirror~=1.1.0
 setuptools~=47.1.0
-pandas~=1.1.4
