+++ docs/paper.bib
@@ -60,7 +60,8 @@
     number = {9},
     pages = {4442-4448},
     year = {2019},
-    month = {04}
+    month = {04},
+    doi = {10.1093/nar/gkz246}
 }
 
 @article{Steenwyk:2019,
@@ -73,7 +74,8 @@ month={Aug},
 day={27},
 volume={12},
 number={1},
-pages={541}
+pages={541},
+doi = {10.1186/s13104-019-4577-5}
 }
 
 @article{Oleary:2015,
@@ -115,6 +117,7 @@ pages={541}
               Sullivan, John P and Sun, Lu and Turner, Se{\'a}n and
               Karsch-Mizrachi, Ilene",
   journal  = "Database (Oxford)",
+  doi      = {10.1093/database/baaa062},
   volume   =  2020,
   month    =  jan,
   year     =  2020
