### 2022-07-09 #4559 @ggonnella
> Also, due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.

OK. Please note that an additional part of the software described in the paper (which synchronises the local copy of the NCBI taxonomy tree) was developed into a different repository (ntmirror), also cited in the paper. This was done since the two parts can be employed also separately.

This is the output of cloc on the ntmirror repository:

Language                     files          blank        comment           code
Python                          18            125            173            720
Text                             8             38              0            164
Markdown                         4             33              0             79
make                             1              8              4             20
SUM:                            31            204            177            983

Thus the total lines of code of Python is currently 1742.

Please let me known if the addition of the second repository under the main repository is really necessary.
