### 2022-07-09 f0946ff
move example application to genomes-attributes-viewer
and make its name and path clear in the paper draft
