+++ docs/paper.bib
@@ -31,7 +31,7 @@
   year = {2022},
   publisher = {Github},
   journal = {Github repository},
-  url = {https://github.com/ggonnella/ntmirror}
+  url = {https://github.com/ggonnella/fastsubtrees/tree/main/ntmirror}
 }
 
 @misc{FastsubtreesRepo,
