### 2022-07-09 #4559 @ggonnella
> Also, due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.

@danielskatz OK. Please note that an additional part of the software described in the paper (which synchronises the local copy of the NCBI taxonomy tree, optionally loads into a database and allows for hierarchical SQL queries, which were used in the paper as a reference benchmark) was developed into a different repository (https://github.com/ggonnella/ntmirror), also cited in the paper. This was done since the two parts can be employed also separately.

To simplify the evaluation the ntmirror sub-package was now moved (by merging the git histories) into a sub-directory of the fastsubtrees repository and this was made clear in the text of the paper.

Thus note that the number of code lines is also higher than counted at the beginning of this thread. For this reason, the lines were re-counted above. The Python LOC are in total 1639 according to the ``cloc`` tool.
