### 2022-07-14 c39deaf
fix specification of package requirements for ntmirror
