+++ requirements.txt
@@ -1,10 +1,1 @@
-docopt~=0.6.2
-loguru~=0.5.3
-tqdm~=4.62.3
-schema~=0.7.4
-sh~=1.14.2
-PyYAML~=6.0
-fastsubtrees~=0.2
-pytest~=7.1.1
-ntmirror~=1.1.0
-setuptools~=47.1.0
+-e .
