### 2022-07-14 dfc1ca0
fix detection of already deleted node
