+++ CHANGELOG.txt
@@ -1,3 +1,7 @@
+=== 1.1 ===
+- fix package requirements specification
+- fix detection of already deleted node
+
 === 1.0 ===
 - version described in the paper
 
