### 2022-08-03 #4559 @arfon
> I could take this on, sure. As a heads up I'm planning to be out of office September 12 - October 3, if it makes sense to ping another editor.

Thanks for the heads up @kellyrowland. Given capacity elsewhere, I would love to have you get this submission moving (under review etc.) and if needs be, I can provide a support if issues arise when you're OoO (please @mention me on the review thread).
