### 2022-08-24 #4559 @vinisalazar
Hello @kellyrowland thank you for the invite. I am interested but my availability is limited in the next few weeks. What would be the expected timeframe?
