### 2022-09-07 #4559 @kellyrowland
Hi @ggonnella @vinisalazar -  I just wanted to let you know that I will be out of office and not checking emails, Github, etc. starting September 10 and will be returning on October 3. Happy to field any questions through the end of the week.

@openjournals/joss-eics would it be possible to loop in another editor here to help find a second reviewer to get this paper to the review stage while I'm out of office?
