### 2022-09-12 #4755 @editorialbot
[REVIEW]: Fastsubtrees: simple and efficient subtrees extractions in Python with applications to NCBI taxonomy
**Submitting author:** @ggonnella (<a href="http://orcid.org/0000-0003-3900-5397">Giorgio Gonnella</a>)
**Repository:** https://github.com/ggonnella/fastsubtrees
**Branch with paper.md** (empty if default branch): main
**Version:** 2.1
