### 2022-09-12 ggonnella/fastsubtrees#3 @KonradHoeffner
pip install fails
Using Python 3.10.7 on Arch Linux, `pip install fastsubtrees` fails with the output below.
This is an acceptance blocker for the JOSS review <https://github.com/openjournals/joss-reviews/issues/4755>.

```
 pip install fastsubtrees
Defaulting to user installation because normal site-packages is not writeable
Collecting fastsubtrees
  Downloading fastsubtrees-1.1-py3-none-any.whl (17 kB)
Collecting tqdm>=4.57.0
  Downloading tqdm-4.64.1-py2.py3-none-any.whl (78 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 78.5/78.5 kB 3.6 MB/s eta 0:00:00
Collecting sh>=1.14.2
  Downloading sh-1.14.3.tar.gz (62 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 62.9/62.9 kB 6.2 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Collecting loguru>=0.5.1
  Downloading loguru-0.6.0-py3-none-any.whl (58 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 58.3/58.3 kB 5.0 MB/s eta 0:00:00
Collecting docopt>=0.6.2
  Downloading docopt-0.6.2.tar.gz (25 kB)
  Preparing metadata (setup.py) ... done
Collecting ntmirror>=1.2
  Downloading ntmirror-1.2-py3-none-any.whl (10 kB)
Collecting schema>=0.7.4
  Downloading schema-0.7.5-py2.py3-none-any.whl (17 kB)
Requirement already satisfied: PyYAML>=6.0 in /usr/lib/python3.10/site-packages (from fastsubtrees) (6.0)
Collecting mariadb
  Downloading mariadb-1.1.4.zip (97 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 97.4/97.4 kB 7.5 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [19 lines of output]
      /bin/sh: line 1: mariadb_config: command not found
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/tmp/pip-install-mj6twek8/mariadb_db9a53888c2a40b99114f01c2c82d527/setup.py", line 27, in <module>
          cfg = get_config(options)
        File "/tmp/pip-install-mj6twek8/mariadb_db9a53888c2a40b99114f01c2c82d527/mariadb_posix.py", line 62, in get_config
          cc_version = mariadb_config(config_prg, "cc_version")
        File "/tmp/pip-install-mj6twek8/mariadb_db9a53888c2a40b99114f01c2c82d527/mariadb_posix.py", line 28, in mariadb_config
          raise EnvironmentError(
      OSError: mariadb_config not found.

      This error typically indicates that MariaDB Connector/C, a dependency which
      must be preinstalled, is not found.
      If MariaDB Connector/C is not installed, see installation instructions
      If MariaDB Connector/C is installed, either set the environment variable
      MARIADB_CONFIG or edit the configuration file 'site.cfg' to set the
       'mariadb_config' option to the file location of the mariadb_config utility.

      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
