### 2022-09-12 ggonnella/fastsubtrees#4 @KonradHoeffner
tests fail
Using Python 3.10.7 on Arch Linux, `pytest` fails with the output below.
This is an acceptance blocker for the JOSS review https://github.com/openjournals/joss-reviews/issues/4755.

```
fastsubtrees$ pytest
==================================================================== test session starts =====================================================================
rootdir: /home/konrad/tmp/fastsubtrees
collected 10 items / 1 error

=========================================================================== ERRORS ===========================================================================
________________________________________________________ ERROR collecting tests/test_fastsubtrees.py _________________________________________________________
ImportError while importing test module '/home/konrad/tmp/fastsubtrees/tests/test_fastsubtrees.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
tests/test_fastsubtrees.py:6: in <module>
    from tests.reference_results import *
E   ModuleNotFoundError: No module named 'tests'
====================================================================== warnings summary ======================================================================
ntmirror/tests/test_dbload_cli.py:8
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_dbload_cli.py:8: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_dbload_cli.py:21
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_dbload_cli.py:21: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:20
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:20: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:33
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:33: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:46
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:46: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
================================================================== short test summary info ===================================================================
ERROR tests/test_fastsubtrees.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================================ 5 warnings, 1 error in 0.17s ================================================================
```
