### 2022-09-12 #4755 @KonradHoeffner
## Functionality
### Installation
Installation only works with an undocumented workaround, see Documentation -> installation instructions
### Performance
I tried to run the fastsubtrees-construct script but it is not clear how exactly to reproduce the tables in the paper.
There should be a single parameterless script that runs all the benchmarks.
A Dockerfile that installs all necessary python and other dependencies would probably be especially helpful, as it would get around the platform specific dependencies.

## Documentation

### Installation instructions
* installation instructions mention `pip install fastsubtrees` but that fails, see https://github.com/ggonnella/fastsubtrees/issues/3. `setup.py` includes the required dependencies, except for some apparently missing MariaDB component.
 A workaround, at least on Arch Linux, is to install `https://aur.archlinux.org/packages/python-mariadb-connector` however this is not mentioned in the documentation.

### Automated tests
There is a test folder with pyttest test scripts but there the tests fail to run, see <https://github.com/ggonnella/fastsubtrees/issues/4>.

### Community guidelines
There are no explicit community guidelines in the README.md but it's a GitHub repository with an issue tracker for reporting issues and seek support, but the way the checklist is written implies this is not enough

## Other

* typo: "Aknownledgements"
