### 2022-09-12 ggonnella/fastsubtrees#4 @KonradHoeffner
The first results occur after `pip install fastsubtrees`. When using `make install` instead the following errors occur:

```
fastsubtrees$ pytest
==================================================================== test session starts =====================================================================
rootdir: /home/konrad/tmp/fastsubtrees
collected 29 items

ntmirror/tests/test_dbload_api.py EE                                                                                                                   [  6%]
ntmirror/tests/test_dbload_cli.py EE                                                                                                                   [ 13%]
ntmirror/tests/test_downloader_api.py ...                                                                                                              [ 24%]
ntmirror/tests/test_downloader_cli.py EEE                                                                                                              [ 34%]
tests/test_fastsubtrees.py ...................                                                                                                         [100%]

=========================================================================== ERRORS ===========================================================================
__________________________________________________________ ERROR at setup of test_dbload_api_mysql ___________________________________________________________

    @pytest.fixture(scope="session")
    def connection_string():
>     config = get_config()

ntmirror/tests/conftest.py:40:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def get_config():
      config_file_path = os.path.join(os.path.dirname(__file__), 'config.yaml')
      if not os.path.exists(config_file_path):
>       raise FileNotFoundError(\
            'Connection configuration file config.yaml not found')
E       FileNotFoundError: Connection configuration file config.yaml not found

ntmirror/tests/conftest.py:32: FileNotFoundError
________________________________________________________ ERROR at setup of test_dbload_api_sqlalchemy ________________________________________________________

    @pytest.fixture(scope="session")
    def connection_string():
>     config = get_config()

ntmirror/tests/conftest.py:40:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def get_config():
      config_file_path = os.path.join(os.path.dirname(__file__), 'config.yaml')
      if not os.path.exists(config_file_path):
>       raise FileNotFoundError(\
            'Connection configuration file config.yaml not found')
E       FileNotFoundError: Connection configuration file config.yaml not found

ntmirror/tests/conftest.py:32: FileNotFoundError
__________________________________________________________ ERROR at setup of test_dbload_cli_mysql ___________________________________________________________

    @pytest.fixture(scope="session")
    def connection_string():
>     config = get_config()

ntmirror/tests/conftest.py:40:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def get_config():
      config_file_path = os.path.join(os.path.dirname(__file__), 'config.yaml')
      if not os.path.exists(config_file_path):
>       raise FileNotFoundError(\
            'Connection configuration file config.yaml not found')
E       FileNotFoundError: Connection configuration file config.yaml not found

ntmirror/tests/conftest.py:32: FileNotFoundError
________________________________________________________ ERROR at setup of test_dbload_cli_sqlalchemy ________________________________________________________

    @pytest.fixture(scope="session")
    def connection_string():
>     config = get_config()

ntmirror/tests/conftest.py:40:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def get_config():
      config_file_path = os.path.join(os.path.dirname(__file__), 'config.yaml')
      if not os.path.exists(config_file_path):
>       raise FileNotFoundError(\
            'Connection configuration file config.yaml not found')
E       FileNotFoundError: Connection configuration file config.yaml not found

ntmirror/tests/conftest.py:32: FileNotFoundError
___________________________________________________________ ERROR at setup of test_first_download ____________________________________________________________
file /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py, line 20
  @pytest.mark.script_launch_mode('subprocess')
  def test_first_download(testout, script, script_runner):
E       fixture 'script_runner' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, connection, connection_args, connection_string, doctest_namespace, monkeypatch, mysql_connection_data, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, script, testdata, testdatadir, testout, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

/home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:20
__________________________________________________________ ERROR at setup of test_no_newer_version ___________________________________________________________
file /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py, line 33
  @pytest.mark.script_launch_mode('subprocess')
  def test_no_newer_version(testout, script, script_runner):
E       fixture 'script_runner' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, connection, connection_args, connection_string, doctest_namespace, monkeypatch, mysql_connection_data, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, script, testdata, testdatadir, testout, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

/home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:33
____________________________________________________________ ERROR at setup of test_newer_version ____________________________________________________________
file /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py, line 46
  @pytest.mark.script_launch_mode('subprocess')
  def test_newer_version(testout, script, script_runner):
E       fixture 'script_runner' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, connection, connection_args, connection_string, doctest_namespace, monkeypatch, mysql_connection_data, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, script, testdata, testdatadir, testout, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

/home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:46
====================================================================== warnings summary ======================================================================
ntmirror/tests/test_dbload_cli.py:8
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_dbload_cli.py:8: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_dbload_cli.py:21
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_dbload_cli.py:21: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:20
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:20: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:33
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:33: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

ntmirror/tests/test_downloader_cli.py:46
  /home/konrad/tmp/fastsubtrees/ntmirror/tests/test_downloader_cli.py:46: PytestUnknownMarkWarning: Unknown pytest.mark.script_launch_mode - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.script_launch_mode('subprocess')

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
================================================================== short test summary info ===================================================================
ERROR ntmirror/tests/test_dbload_api.py::test_dbload_api_mysql - FileNotFoundError: Connection configuration file config.yaml not found
ERROR ntmirror/tests/test_dbload_api.py::test_dbload_api_sqlalchemy - FileNotFoundError: Connection configuration file config.yaml not found
ERROR ntmirror/tests/test_dbload_cli.py::test_dbload_cli_mysql - FileNotFoundError: Connection configuration file config.yaml not found
ERROR ntmirror/tests/test_dbload_cli.py::test_dbload_cli_sqlalchemy - FileNotFoundError: Connection configuration file config.yaml not found
ERROR ntmirror/tests/test_downloader_cli.py::test_first_download
ERROR ntmirror/tests/test_downloader_cli.py::test_no_newer_version
ERROR ntmirror/tests/test_downloader_cli.py::test_newer_version
========================================================== 22 passed, 5 warnings, 7 errors in 7.31s ==========================================================
```
