### 2022-09-26 #4755 @KonradHoeffner
@kellyrowland: I am finished with my first review and checked all the boxes I feel comfortable checking.
Unfortunately I cannot accept the paper at the current state due to the reasons commented above.
However if the authors fix the two acceptance blockers (issues filed at the target repository because of failing installation and failing tests) and address the other comments, this can be changed in my opinion.
As this is my first review, I'm not sure how I signal to the editorialbot that I am finished reviewing, there does not seem to be an editorialbot command where I can submit my recommendation, so I hope just writing this comment is enough.
