### 2022-10-02 59cead2
use fixtures instead of reference_results module in tests
