### 2022-10-02 3f61193
use fixtures and cleanup test suite
