+++ tests/reference_results.py
@@ -1,11 +1,0 @@
-from array import array
-
-results_query_smalltree_id_1 = array('Q', [1, 2, 8, 3, 7, 9, 4, 5])
-results_query_smalltree_id_8 = array('Q', [8, 3, 7, 9, 4, 5])
-results_query_middletree_id_1 = array('Q', [1, 2, 4, 14, 58, 66, 162, 849, 170, 426, 769, 772, 243, 308, 395, 564, 263, 510, 167, 444, 472, 985, 452, 659, 527, 812, 940, 968, 307, 374, 983, 18, 19, 38, 83, 256, 572, 601, 996, 261, 323, 333, 945, 463, 509, 638, 838, 349, 922, 41, 492, 75, 474, 724, 373, 576, 21, 55, 181, 280, 516, 591, 926, 963, 583, 759, 814, 93, 130, 160, 220, 361, 404, 835, 455, 700, 773, 976, 920, 369, 850, 427, 525, 582, 625, 851, 861, 165, 189, 811, 346, 603, 696, 405, 860, 916, 420, 691, 276, 381, 592, 598, 133, 156, 250, 476, 880, 300, 530, 543, 557, 575, 33, 269, 397, 942, 46, 78, 110, 663, 939, 836, 955, 881, 235, 358, 451, 774, 755, 914, 91, 367, 549, 164, 974, 882, 270, 443, 632, 671, 664, 767, 800, 6, 67, 222, 281, 448, 697, 347, 847, 497, 137, 431, 522, 658, 640, 688, 900, 253, 259, 324, 362, 363, 641, 843, 845, 868, 783, 344, 887, 16, 188, 20, 60, 825, 62, 65, 81, 106, 116, 123, 217, 260, 356, 820, 757, 842, 441, 534, 645, 719, 971, 801, 247, 478, 844, 877, 602, 173, 617, 686, 141, 513, 190, 681, 490, 311, 588, 639, 864, 994, 329, 853, 736, 29, 45, 339, 350, 578, 949, 68, 99, 180, 262, 317, 377, 435, 826, 413, 682, 597, 765, 351, 703, 931, 203, 506, 507, 552, 600, 986, 97, 199, 213, 440, 743, 791, 519, 134, 150, 186, 352, 417, 871, 569, 782, 241, 550, 810, 271, 442, 540, 752, 545, 637, 828, 336, 372, 387, 400, 496, 512, 713, 822, 623, 355, 562, 630, 902, 888, 655, 957, 676, 746, 927, 804, 3, 5, 17, 25, 64, 73, 118, 305, 401, 321, 982, 456, 695, 803, 948, 331, 559, 965, 749, 147, 342, 571, 806, 763, 185, 567, 993, 187, 786, 950, 230, 244, 689, 699, 701, 407, 23, 192, 503, 643, 799, 660, 28, 47, 59, 148, 201, 627, 214, 670, 834, 967, 1000, 466, 61, 288, 411, 518, 526, 788, 648, 103, 174, 299, 535, 766, 194, 393, 863, 903, 475, 568, 590, 650, 925, 207, 998, 277, 915, 306, 462, 969, 191, 325, 768, 398, 445, 524, 809, 502, 548, 787, 12, 13, 24, 135, 139, 539, 705, 414, 556, 716, 35, 40, 127, 193, 228, 234, 382, 684, 290, 368, 565, 318, 493, 960, 131, 211, 704, 152, 479, 657, 958, 744, 777, 100, 365, 396, 459, 43, 69, 340, 873, 345, 500, 467, 89, 126, 107, 314, 930, 119, 236, 285, 423, 460, 508, 636, 515, 245, 434, 718, 771, 388, 453, 410, 730, 886, 946
[…]

[…]
