### 2022-10-02 a54a8e6
ntmirror: db tests only if db config file is available
