### 2022-10-02 9f1bb93
add snacli to ntmirror requirements
