### 2022-10-02 d62581a
tests run correctly in Docker container
(skipping ntmirror database tests)
