### 2022-10-03 e2e0d22
make also db-dependent tests run in docker image
