+++ docker/ntmirror.config.yaml
@@ -1,0 +1,7 @@
+database: ntmirror_test
+username: myuser
+password: mypass
+host: localhost
+port: 3306
+socket: /run/mysqld/mysqld.sock
+drivername: "mysql+mysqldb"
