### 2022-10-03 9627d72
option --reset for ntmirror-dbload
it allows easier benchmarking of dbload
