+++ docker/benchmarks_sql.sh
@@ -1,2 +1,2 @@
-time ntmirror-dbload myuser mypass ntmirror_test /run/mysqld/mysqld.sock ntdumpdir
+time ntmirror-dbload --reset myuser mypass ntmirror_test /run/mysqld/mysqld.sock ntdumpdir
 
