### 2022-10-03 3e9b006
more work on benchmarks under docker
