+++ docker/benchmarks_sql.sh
@@ -1,2 +1,10 @@
-time ntmirror-dbload --reset myuser mypass ntmirror_test /run/mysqld/mysqld.sock ntdumpdir
+#!/bin/bash
+
+NREPEATS=3
+
+for ((i=0; i<$NREPEATS; i++)); do
+    echo "Iteration $i"
+    /usr/bin/time ntmirror-dbload --testmode --reset \
+      myuser mypass ntmirror_test /run/mysqld/mysqld.sock ntdumpdir
+done
 
