### 2022-10-03 1fbb0c2
more work on benchmarks in docker
