### 2022-10-03 6ad634d
continue implementation of benchmarks in Docker image
