### 2022-10-03 8af88be
add schema to ntmirror req libraries
