+++ ntmirror/setup.py
@@ -17,7 +17,7 @@ setup(name='ntmirror',
       long_description=readme(),
       long_description_content_type="text/markdown",
       install_requires=["sqlalchemy", "sqlalchemy_repr", "mysql", "mariadb",
-        "loguru", "snacli"],
+        "loguru", "snacli", "schema"],
       url='https://github.com/ggonnella/ntmirror',
       keywords="ncbi taxonomy database mirror",
       author='Giorgio Gonnella',
