### 2022-10-03 d21e327
fix benchmarks_fst 2
