+++ docker/Dockerfile
@@ -19,8 +19,7 @@ RUN apt-get install -y libmariadb3 libmariadb-dev
 RUN pip install mariadb pytest PyYAML sqlalchemy sh pytest-console-scripts
 RUN pip install schema snacli ntmirror
 RUN ntmirror-download ntdumpdir
-ARG VER=unknown
-RUN VER=${VER} git clone -b main https://github.com/ggonnella/fastsubtrees.git /fastsubtrees
+RUN DATE=`date` git clone -b main https://github.com/ggonnella/fastsubtrees.git /fastsubtrees
 RUN pip uninstall ntmirror -y && cd fastsubtrees/ntmirror && pip install -e .
 COPY ./ntmirror.config.yaml /fastsubtrees/ntmirror/tests/config.yaml
 RUN cd fastsubtrees && pip install -e .
