+++ ntmirror/setup.py
@@ -38,7 +38,8 @@ setup(name='ntmirror',
       ],
       packages=find_packages(),
       scripts=['bin/ntmirror-download',
-               'bin/ntmirror-dbload'],
+               'bin/ntmirror-dbload',
+               'bin/ntmirror-extract-subtree'],
       zip_safe=False,
       test_suite="pytest",
       include_package_data=True,
