### 2022-10-04 7b885ea
avoid unnecessary instance var
