### 2022-10-04 ca40a56
genomes-attribute-viewer: fix Werkzeug version
