### 2022-10-04 881fd78
repair generate_attribute_file script
