### 2022-10-04 49b0b54
easier setup and run of genomes-attributes-viewer
