### 2022-10-04 df3beca
genome-attributes-viewer app from inside Docker
