### 2022-10-04 02827d2
rm container before image
