+++ genomes-attributes-viewer/requirements.txt
@@ -1,6 +1,6 @@
 ntmirror
 dash==2.0.0
-Werkzeug=2.0.0
+Werkzeug==2.0.0
 dash-bootstrap-components==1.0.2
 dash-core-components==2.0.0
 dash-html-components==2.0.0
