### 2022-10-04 ce21b7f
fix paths in Dockerfile
