### 2022-10-04 19ccac0
fix Flask version for genome-attributes-viewer
