### 2022-10-04 2cdf082
start_app for starting genome-attributes-viewer in Docker container
