### 2022-10-04 6cf3511
add pandas to genome-attributes-viewer requirements
