+++ genomes-attributes-viewer/requirements.txt
@@ -6,3 +6,4 @@ dash-html-components==2.0.0
 dash-table==5.0.0
 Werkzeug==2.0.0
 Flask==2.1.2
+pandas
