### 2022-10-04 d779133
genomes-attributes-viewer setup, skip download if ntdumps dir exists
(useful to speedup Docker image development)
