### 2022-10-04 f43b8a3
Dockerfile, re-use nt.tree file
