### 2022-10-05 da5847c
use 0.0.0.0 as IP for the genomes-attributes-viewer app
