### 2022-10-05 f7878e3
example application now works correctly in Docker container
