### 2022-10-05 40bb21e
move provided example attribute data into own directory
