+++ data/README.md
@@ -1,0 +1,5 @@
+The table in the file ``accession_taxid_attribute.tsv.gz`` contains
+the values of some sequence statistics computed for bacterial genomes,
+downloaded on May 28th, 2022 from NCBI Refseq
+(ftp.ncbi.nlm.nih.gov/genomes/refseq/bacteria/),
+using the script ``fas_stats_posix.py``.
