+++ data/README.md
@@ -3,3 +3,6 @@ the values of some sequence statistics computed for bacterial genomes,
 downloaded on May 28th, 2022 from NCBI Refseq
 (ftp.ncbi.nlm.nih.gov/genomes/refseq/bacteria/),
 using the script ``fas_stats_posix.py``.
+
+The file is used for benchmarks and for the example application
+genomes-attributes-viewer.
