### 2022-10-05 a7bf2a9
remove unnecessary argument in fastsubtrees-attribtues-construct script
