### 2022-10-05 0654fa8
fix fastsubtrees-attributes-construct and provide an example module
