### 2022-10-05 ae85777
fix benchmarks_attr script
