+++ docker/benchmarks_attr.sh
@@ -13,7 +13,7 @@ for attr in genome_size GC_content; do
   for ((i=0; i<$NREPEATS; i++)); do
     STEP="construct-$attr"
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-      fastsubtrees-attributes-construct $attr.attr /ncbi-taxonomy.tree
+      fastsubtrees-attributes-construct $attr.attr /ncbi-taxonomy.tree \
         /fastsubtrees/data/attribute_values.py --keyargs \
           filename=/fastsubtrees/data/accession_taxid_attribute.tsv.gz \
           taxid_col=1 attr_col=${attr_col[$attr]}
@@ -25,7 +25,7 @@ for attr in genome_size GC_content; do
     for ROOT in 511145 83333 562 561 543 91347 1236 1224 2; do
     echo "Step $STEP from node $ROOT, iteration $i..."
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-      fastsubtrees-attribute-query
+      fastsubtrees-attribute-query \
         /nbci-taxonomy.tree $ROOT $attr.attr > \
       attr_values.$attr.$ROOT
     done
