### 2022-10-05 f6e67df
fix again benchmarks_attr script
