### 2022-10-05 40b07d5
specify datatype of attributes in benchmark script
