### 2022-10-05 dadea6d
fix in benchmarks_attr script
