### 2022-10-05 bbd1f76
final version of benchmarks_attr script
