+++ docker/benchmarks_attr.sh
@@ -24,13 +24,12 @@ for attr in genome_size GC_content; do
             taxid_col=1 attr_col=${attr_col[$attr]}
   done
 
-  rm -f $OUTFILE
   for ((i=0; i<$NREPEATS; i++)); do
     STEP="query-$attr"
     for ROOT in 511145 83333 562 561 543 91347 1236 1224 2; do
     echo "Step $STEP from node $ROOT, iteration $i..."
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-      fastsubtrees-attributes-query \
+      fastsubtrees-attributes-query --filter --countN --countV \
         /ncbi-taxonomy.tree $ROOT $attr.attr > \
       attr_values.$attr.$ROOT
     done
