+++ benchmarks/benchmarks_attr.sh
@@ -1,9 +1,27 @@
 #!/bin/bash
 
+#
+# Benchmark the construction and query of the attribute files
+# associated with a tree.
+#
+
 NREPEATS=3
 OUTFILE=benchmarks_attr.tsv
 
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 <tree>"
+    echo "where <tree> is the name of a fastsubtrees tree file constructed"
+    echo "with fastsubtrees-construct from a NCBI taxonomy dump"
+    exit 1
+fi
+TREE=$1
+
 rm -f $OUTFILE
+mkdir -p results
+
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+FST_DIR=$SCRIPT_DIR/..
+FST_DATA_DIR=$FST_DIR/data
 
 declare -A attr_col
 attr_col[genome_size]=2
@@ -17,10 +35,10 @@ for attr in genome_size GC_content; do
   for ((i=0; i<$NREPEATS; i++)); do
     STEP="construct-$attr"
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-      fastsubtrees-attributes-construct $attr.attr /ncbi-taxonomy.tree \
-        /fastsubtrees/data/attribute_values.py --datatype ${attr_dt[$attr]} \
+      fastsubtrees-attributes-construct $attr.attr $TREE \
+        $FST_DATA_DIR/attribute_values.py --datatype ${attr_dt[$attr]} \
           --keyargs \
-            filename=/fastsubtrees/data/accession_taxid_attribute.tsv.gz \
+            filename=$FST_DATA_DIR/accession_taxid_attribute.tsv.gz \
             taxid_col=1 attr_col=${attr_col[$attr]}
   done
 
@@ -30,8 +48,8 @@ for attr in genome_size GC_content; do
     echo "Step $STEP from node $ROOT, iteration $i..."
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
       fastsubtrees-attributes-query --filter --countN --countV \
-        /ncbi-taxonomy.tree $ROOT $attr.attr > \
-      attr_values.$attr.$ROOT
+        $TREE $ROOT $attr.attr > \
+      results/attr_values.$attr.$ROOT
     done
   done
 done

[…]
