### 2022-10-05 71590a8
split fst construct and query in benchmarks
