### 2022-10-05 e4aa2bd
update paths in Makefile
