+++ docker/Makefile
@@ -43,18 +43,18 @@ test: create_test_db
 
 benchmarks_sql.tsv: create_test_db
 	docker exec -it ${CONTAINER_NAME} bash \
-		-e /fastsubtrees/docker/benchmarks_sql.sh \
+		-e /fastsubtrees/benchmarks/benchmarks_sql.sh \
 	  	myuser mypass ntmirror_test /run/mysqld/mysqld.sock
 	docker cp ${CONTAINER_NAME}:/benchmarks_sql.tsv .
 
 benchmarks_contruct.tsv: container
 	docker exec -it ${CONTAINER_NAME} bash \
-		-e /fastsubtrees/docker/benchmarks_construct.sh /ntdumpdir
+		-e /fastsubtrees/benchmarks/benchmarks_construct.sh /ntdumpdir
 	docker cp ${CONTAINER_NAME}:/benchmarks_construct.tsv .
 
 benchmarks_fst.tsv: container
 	docker exec -it ${CONTAINER_NAME} bash \
-		-e /fastsubtrees/docker/benchmarks_fst.sh /ncbi-taxonomy.tree
+		-e /fastsubtrees/benchmarks/benchmarks_fst.sh /ncbi-taxonomy.tree
 	docker cp ${CONTAINER_NAME}:/benchmarks_fst.tsv .
 
 benchmarks_attr.tsv: container
