+++ docker/Makefile
@@ -47,7 +47,7 @@ benchmarks_sql.tsv: create_test_db
 	  	myuser mypass ntmirror_test /run/mysqld/mysqld.sock
 	docker cp ${CONTAINER_NAME}:/benchmarks_sql.tsv .
 
-benchmarks_contruct.tsv: container
+benchmarks_construct.tsv: container
 	docker exec -it ${CONTAINER_NAME} bash \
 		-e /fastsubtrees/benchmarks/benchmarks_construct.sh /ntdumpdir
 	docker cp ${CONTAINER_NAME}:/benchmarks_construct.tsv .
@@ -62,6 +62,7 @@ benchmarks_attr.tsv: container
 		-e /fastsubtrees/benchmarks/benchmarks_attr.sh /ncbi-taxonomy.tree
 	docker cp ${CONTAINER_NAME}:/benchmarks_attr.tsv .
 
-benchmarks: benchmarks_sql.tsv benchmarks_fst.tsv benchmarks_attr.tsv
+#benchmarks: benchmarks_sql.tsv benchmarks_fst.tsv benchmarks_attr.tsv
+benchmarks: benchmarks_fst.tsv benchmarks_attr.tsv
 
 benchmarks_all: benchmarks benchmarks_construct.tsv
