### 2022-10-05 35f3635
define common benchmark parameters in own file
