+++ benchmarks/benchmarks_params.sh
@@ -1,0 +1,4 @@
+export NREPEATS=3
+export NODES="511145 83333 562 561 543 91347 1236 1224 2"
+export OUTDIR="results"
+export OUTFILE_PFX="benchmarks_"
