+++ docker/Makefile
@@ -62,7 +62,6 @@ benchmarks_attr.tsv: container
 		-e /fastsubtrees/benchmarks/benchmarks_attr.sh /ncbi-taxonomy.tree
 	docker cp ${CONTAINER_NAME}:/benchmarks_attr.tsv .
 
-#benchmarks: benchmarks_sql.tsv benchmarks_fst.tsv benchmarks_attr.tsv
-benchmarks: benchmarks_fst.tsv benchmarks_attr.tsv
+benchmarks: benchmarks_sql.tsv benchmarks_fst.tsv benchmarks_attr.tsv
 
 benchmarks_all: benchmarks benchmarks_construct.tsv
