+++ ntmirror/CHANGELOG.txt
@@ -1,3 +1,9 @@
+== 1.3.0 ==
+- version number changed to avoid a version name clash
+
+== 1.2.0 ==
+- fixes and improvements, see fastsubtrees v1.2 changelog
+
 == 1.1.0 ==
 - add ntmirror-extract-subtree script
 
