### 2022-10-06 559d3ae
add make goal to download image from DockerHub
