+++ CHANGELOG.txt
@@ -1,3 +1,6 @@
+=== 1.2.1 ===
+- make sure ids_modules is included in the package by adding a __init__.py
+
 === 1.2 ===
 - docker image providing tests, benchmarks and example app
 - tests divided into multiple files and using fixtures
