### 2022-10-06 bf619f2
simplified handling of attribute files using the CLI
