+++ fastsubtrees/attribute.py
@@ -1,4 +1,17 @@
 import json
+import glob
+
+EXT = "attr"
+
+def attrfilename(treefilename, attribute):
+  return "{}.{}.{}".format(treefilename, attribute, EXT)
+
+def attrfiles(treefilename):
+  result = {}
+  for file in glob.glob(treefilename + ".*." + EXT):
+    attribute = file[len(treefilename)+1:-(len(EXT))-1]
+    result[attribute] = file
+  return result
 
 def get_attribute_value(attributefile, subtree_size, position):
   with open(attributefile, 'r') as file:
@@ -15,3 +28,9 @@ def get_attribute_list(tree, subtree_root, attributefile):
   attribute_list = \
       get_attribute_value(attributefile, subtree_size+1, position-1)
   return attribute_list
+
+def write_attributes_values(tree, attrvalues, outfile):
+  for element_id in tree.subtree_ids(tree.root_id):
+    attribute = attrvalues.get(element_id, None)
+    outfile.write(json.dumps(attribute) + "\n")
+
