+++ setup.py
@@ -36,8 +36,8 @@ setup(name='fastsubtrees',
       packages=find_packages(),
       scripts=['bin/fastsubtrees-construct',
                'bin/fastsubtrees-query',
-               'bin/fastsubtrees-attributes-construct',
-               'bin/fastsubtrees-attributes-query',
+               'bin/fastsubtrees-attr-construct',
+               'bin/fastsubtrees-attr-query',
                'bin/fastsubtrees-add-subtree',
                'bin/fastsubtrees-delete-subtree'],
       zip_safe=False,
