### 2022-10-06 0b6963c
use new conventions for attr filenames in example app
