### 2022-10-06 fd89461
use new attr script names in benchmarks_attr
