### 2022-10-06 7b76452
fix _scripts_support logger calls and remove spurious fn
