+++ fastsubtrees/_scripts_support.py
@@ -15,7 +15,7 @@ def setup_verbosity(args):
     fastsubtrees.enable_logger(level)
 
 def get_module(fname, function):
-  logger.debug("Loading Python module '{}'".format(fname))
+  fastsubtrees.logger.debug("Loading Python module '{}'".format(fname))
   modulename = Path(fname).stem
   spec = importlib.util.spec_from_file_location(modulename, fname)
   m = importlib.util.module_from_spec(spec)
@@ -23,7 +23,8 @@ def get_module(fname, function):
   if not m.__dict__.get(function):
     raise ValueError("The specified Python module {} does not define a "
                      "function {}()".format(fname, function))
-  logger.success("Module loaded, function {}() found".format(function))
+  fastsubtrees.logger.success("Module loaded, function {}() found".\
+      format(function))
   return m
 
 def get_fn_args(has_keyargs, values):
@@ -35,36 +36,30 @@ def get_fn_args(has_keyargs, values):
     keyargs = {}
     posargs = values
   if posargs:
-    logger.debug(f"Positional arguments passed to the function: {posargs}")
+    fastsubtrees.logger.debug(\
+        f"Positional arguments passed to the function: {posargs}")
   if keyargs:
-    logger.debug(f"Keyword arguments passed to the function: {keyargs}")
+    fastsubtrees.logger.debug(\
+        f"Keyword arguments passed to the function: {keyargs}")
   return posargs, keyargs
 
 def get_datatype_casting_fn(value, m, mfilename):
   if not value:
-    logger.debug("Using str for datatype casting (default)")
+    fastsubtrees.logger.debug("Using str for datatype casting (default)")
     cast = str
   else:
     try:
       cast = eval(value)
-      logger.debug(f"Using function '{value}()' for datatype casting")
+      fastsubtrees.logger.debug(\
+          f"Using function '{value}()' for datatype casting")
     except NameError:
       if m.__dict__.get(value):
         cast = m.__dict__[value]
-        logger.debug("Using function {}() from module {} for datatype casting".\
-            format(value, mfilename))
+        fastsubtrees.logger.debug(\
+            "Using function {}() from module {} for datatype casting".\
+              format(value, mfilename))
       else:
         raise ValueError("Invalid datatype casting function '{}'".\
             format(value))
     return cast
-
-def find_attribute_files(treefile):
-  # find all files whose path starts with treefile and end with .attrs
-  treefile = Path(treefile)
-  treefile_dir = treefile.parent
-  treefile_prefix = treefile.stem
-  files = [f for f in treefile_dir.iterdir() if f.stem.startswith(treefile_prefix)]
-  # filter out files that are not attribute files
-  files = [f for f in files if f.suffix in [".csv", ".tsv", ".txt"]]
-  return files
 
