+++ CHANGELOG.txt
@@ -1,3 +1,6 @@
+=== 1.3 ===
+- simplify scripts CLI interface
+
 === 1.2.1 ===
 - make sure ids_modules is included in the package by adding a __init__.py
 
