### 2022-10-07 ggonnella/fastsubtrees#4 @ggonnella
The test were fixed now and run on all platforms I have access to.
