### 2022-10-07 ggonnella/fastsubtrees#3 @ggonnella
The problems for the installation depend on the fact that mariadb requires a Python connector, which must be installed before the package can be installed. MariaDB is not a requirement of fastsubtrees itself, but of the subpackage ntmirror
and it is used for comparison with SQL solutions, e.g. in the benchmarks.

I provided now a Docker-based solution (see docker/Dockerfile) in which all the required components are pre-installed, so that the test, benchmarks and the example application can be easily run without effort.
