### 2022-10-07 8a66609
Replace benchmark results with current ones measured using Docker
