+++ benchmarks/Comparison_SQL_Fastsubtrees_Subtree_Values.txt
@@ -1,10 +1,0 @@
-Subtree ID	Subtree Size	Average CPU Time (System + User) SQL	Average CPU Time (System + User) Fastsubtrees	Average Real Time SQL	Average Real Time Fastsubtrees	Memory Peak SQL	Memory Peak Fastsubtrees
-511145	1	0,34	0,37	0,68	0,74	33184	159340
-83333	36	0,33	0,35	0,67	0,7	33236	159920
-562	3379	0,65	0,36	1,75	0,72	41772	156916
-561	4434	0,79	0,37	2,74	0,74	44052	158940
-543	22580	2,22	0,36	7,61	0,72	90012	157884
-91347	31394	2,97	0,37	9,85	0,74	111576	159396
-1236	122607	10,35	0,4	37,5	0,81	341864	163800
-1224	226821	18,87	0,43	78,4	0,95	605384	165136
-2	532460	43,27	1,06	186,35	1,51	1372496	174780
