+++ benchmarks/Fastsubtrees_benchmark.txt
@@ -1,37 +1,0 @@
-Node ID	User	System	Elapsed	Maxresident	No, of IDs
-511145	0,55	0,18	0,73	159340	1
-511145	0,55	0,19	0,75	155320	1
-511145	0,54	0,19	0,74	158436	1
-Average	0,546666667	0,186666667	0,74	157698,6667	
-83333	0,52	0,17	0,7	157600	36
-83333	0,53	0,17	0,71	158796	36
-83333	0,54	0,15	0,7	159920	36
-Average	0,53	0,163333333	0,703333333	158772	
-562	0,56	0,16	0,73	156916	3379
-562	0,53	0,17	0,7	155612	3379
-562	0,55	0,17	0,73	155844	3379
-Average	0,546666667	0,166666667	0,72	156124	
-561	0,56	0,18	0,74	158940	4434
-561	0,56	0,17	0,74	158644	4434
-561	0,56	0,18	0,74	157332	4434
-Average	0,56	0,176666667	0,74	158305,3333	
-543	0,54	0,18	0,72	157884	22580
-543	0,52	0,18	0,71	156812	22580
-543	0,55	0,16	0,72	156068	22580
-Average	0,536666667	0,173333333	0,716666667	156921,3333	
-91347	0,57	0,17	0,75	158092	31394
-91347	0,58	0,17	0,76	159396	31394
-91347	0,54	0,17	0,72	157104	31394
-Average	0,563333333	0,17	0,743333333	158197,3333	
-1236	0,62	0,18	0,82	160692	122607
-1236	0,62	0,19	0,83	163800	122607
-1236	0,59	0,17	0,78	159116	122607
-Average	0,61	0,18	0,81	161202,6667	
-1224	0,66	0,2	0,95	162444	226821
-1224	0,66	0,2	0,95	163512	226821
-1224	0,66	0,21	0,95	165136	226821
-Average	0,66	0,203333333	0,95	163697,3333	
-2	0,81	0,26	1,72	172880	532460
-2	0,83	0,24	1,45	174780	532460
-2	0,84	0,23	1,36	172556	532460
-Average	0,826666667	0,243333333	1,51	173405,3333	
