+++ benchmarks/Fastsubtrees_Attribute_Values_benchmarks.txt
@@ -1,10 +1,0 @@
-Subtree ID	Subtree Size	Average CPU Time (System + User)	Average Real Time	Memory Peak 	No. of Nodes with values	No. of values
-511145	1	0,55	1,11	321952	1	9
-83333	36	0,55	1,1	319604	8	38
-562	3379	0,56	1,16	318804	165	2160
-561	4464	0,56	1,12	321520	174	2246	
-543	22580	0,66	1,33	321116	839	5774
-91347	31394	0,68	1,38	320444	1150	6709
-1236	122607	1	2,01	323028	2819	11167
-1224	226821	1,22	2,73	322664	5090	16063
-2	532460	2	4,35	324052	10043	27515
