+++ benchmarks/SQL_benchmark.txt
@@ -1,37 +1,0 @@
-Node ID	User	System	Elapsed	Maxresident	No, of IDs
-511145	0,57	0,08	0,66	33184	1
-511145	0,59	0,1	0,69	33176	1
-511145	0,61	0,08	0,7	33180	1
-Average	0,59	0,09	0,68	33180	
-83333	0,6	0,07	0,68	33228	36
-83333	0,58	0,08	0,68	33236	36
-83333	0,57	0,06	0,65	33236	36
-Average	0,583333333	0,07	0,67	33233,33333	
-562	0,96	0,32	1,8	41772	3379
-562	0,99	0,32	1,71	41772	3379
-562	0,96	0,37	1,74	41772	3379
-Average	0,97	0,336666667	1,75	41772	
-561	1,19	0,48	4,13	44052	4434
-561	1,12	0,45	2,08	44052	4434
-561	1,08	0,41	2,01	44052	4434
-Average	1,13	0,446666667	2,74	44052	
-543	3	1,53	8,44	90012	22580
-543	2,85	1,58	7,24	90012	22580
-543	2,86	1,52	7,15	90012	22580
-Average	2,903333333	1,543333333	7,61	90012	
-91347	4,2	2,04	10,39	111572	31394
-91347	3,8	2,13	9,8	111568	31394
-91347	3,67	1,97	9,36	111576	31394
-Average	3,89	2,046666667	9,85	111572	
-1236	13	7,6	41,79	341856	122607
-1236	13,04	7,63	35,55	341864	122607
-1236	13,25	7,55	35,15	341732	122607
-Average	13,09666667	7,593333333	37,49666667	341817,3333	
-1224	23,36	14,19	82,32	605336	226821
-1224	24,82	13,69	75,62	605384	226821
-1224	24,41	12,7	77,33	605380	226821
-Average	24,19666667	13,52666667	78,42333333		
-2	54,11	32,38	187,58	1372448	532460
-2	57,11	29,28	185,56	1372496	532460
-2	56,3	30,42	185,91	1372448	532460
-Average	55,84	30,69	186,35	1372464	
