+++ benchmarks/Fastsubtrees_Attribute_GenomeSize.txt
@@ -1,37 +1,0 @@
-Node ID	User	System	Elapsed	Maxresident	Nodes with Attributes	Number of attribute values
-511145	0,79	0,32	1,12	321952	1	9
-511145	0,8	0,3	1,11	317740	1	9
-511145	0,82	0,29	1,11	317620	1	9
-Average	0,803333333	0,303333333	1,113333333	319104		
-83333	0,78	0,32	1,11	319308	8	38
-83333	0,79	0,31	1,11	319604	8	38
-83333	0,78	0,32	1,1	317760	8	38
-Average	0,783333333	0,316666667	1,106666667	318890,6667		
-562	0,79	0,34	1,13	318804	165	2160
-562	0,79	0,3	1,1	318380	165	2160
-562	0,8	0,31	1,12	318640	165	2160
-Average	0,793333333	0,316666667	1,116666667	318608		
-561	0,8	0,32	1,12	321520	174	2246
-561	0,8	0,31	1,12	318596	174	2246
-561	0,81	0,32	1,13	318160	174	2246
-Average	0,803333333	0,316666667	1,123333333	319425,3333		
-543	1,02	0,3	1,33	319108	839	5774
-543	1	0,31	1,32	319092	839	5774
-543	1,04	0,29	1,34	321116	839	5774
-Average	1,02	0,3	1,33	319772		
-91347	1,07	0,3	1,39	319688	1150	6709
-91347	1,05	0,29	1,36	319660	1150	6709
-91347	1,07	0,32	1,4	320444	1150	6709
-Average	1,063333333	0,303333333	1,383333333	319930,6667		
-1236	1,51	0,42	1,95	319076	2819	11167
-1236	1,65	0,37	2,04	320644	2819	11167
-1236	1,67	0,37	2,06	323028	2819	11167
-Average	1,61	0,386666667	2,016666667	320916		
-1224	2,06	0,4	2,69	321360	5090	16063
-1224	2,08	0,37	2,81	320120	5090	16063
-1224	2,09	0,38	2,7	322664	5090	16063
-Average	2,076666667	0,383333333	2,733333333	321381,3333		
-2	3,53	0,47	4,26	321320	10043	27515
-2	3,58	0,49	4,6	323272	10043	27515
-2	3,47	0,48	4,19	324052	10043	27515
-Average	3,526666667	0,48	4,35	322881,3333		

[…]
