### 2022-10-07 035060f
script for extracting tree sizes from output of benchmark_attr.sh
