### 2022-10-07 7d82ffe
script to reproduce table 1 of the paper
