### 2022-10-07 a0729dc
script to reproduce table 2 of the paper
