+++ benchmarks/make_table1.py
@@ -73,12 +73,12 @@ def main(args):
       lines_order.append(subtree_id)
       subtree_size = fields[1]
       results[subtree_id]['subtree_size'] = subtree_size
-  print('|' + ' | '.join([headers["subtree_id"]] + \
-                         [headers[x] for x in order]) + '|')
+  print('| ' + ' | '.join([headers["subtree_id"]] + \
+                         [headers[x] for x in order]) + ' |')
   print('|-'*(len(order)+1) + '|')
   for subtree_id in lines_order:
-    print('|' + ' | '.join([subtree_id] + \
-                           [str(results[subtree_id][x]) for x in order]) + '|')
+    print('| ' + ' | '.join([subtree_id] + \
+                           [str(results[subtree_id][x]) for x in order]) + ' |')
 
 if __name__ == "__main__":
   args = docopt(__doc__)
