+++ benchmarks/table1.md
@@ -1,0 +1,11 @@
+| Subtree root ID | Subtree size | SQL CPU time (s) | fastsubtrees CPU time (s) | SQL real time (s) | fastsubtrees real time (s) | SQL memory peak (MB) | fastsubtrees memory peak (MB) |
+|-|-|-|-|-|-|-|-|
+| 511145 | 1 | 0.15 | 0.13 | 1.39 | 0.13 | 33.8 | 146.3 |
+| 83333 | 36 | 0.15 | 0.12 | 2.62 | 0.12 | 33.9 | 146.3 |
+| 562 | 3380 | 0.17 | 0.12 | 4.84 | 0.12 | 40.8 | 146.2 |
+| 561 | 4435 | 0.17 | 0.12 | 6.15 | 0.12 | 42.6 | 146.3 |
+| 543 | 22747 | 0.37 | 0.12 | 36.80 | 0.13 | 81.0 | 146.2 |
+| 91347 | 31606 | 0.44 | 0.13 | 36.89 | 0.13 | 98.8 | 146.2 |
+| 1236 | 123293 | 1.50 | 0.15 | 60.85 | 0.15 | 289.0 | 146.3 |
+| 1224 | 228115 | 2.57 | 0.18 | 90.96 | 0.18 | 507.2 | 146.3 |
+| 2 | 535205 | 5.87 | 0.24 | 131.03 | 0.24 | 1140.4 | 146.3 |
