+++ benchmarks/table2.md
@@ -1,0 +1,11 @@
+| Subtree root ID | Subtree size | CPU time (s) | Real time (s) | Memory peak (MB) | N. nodes with values | N. values |
+|-|-|-|-|-|-|-|
+| 511145 | 1 | 0.21 | 0.22 | 305.7 | 1 | 9 |
+| 83333 | 36 | 0.21 | 0.22 | 305.7 | 8 | 38 |
+| 562 | 3380 | 0.22 | 0.22 | 306.0 | 165 | 2160 |
+| 561 | 4435 | 0.22 | 0.22 | 305.9 | 174 | 2246 |
+| 543 | 22747 | 0.24 | 0.25 | 306.0 | 839 | 5774 |
+| 91347 | 31606 | 0.25 | 0.25 | 306.0 | 1150 | 6709 |
+| 1236 | 123293 | 0.34 | 0.34 | 306.5 | 2830 | 11178 |
+| 1224 | 228115 | 0.41 | 0.42 | 306.9 | 5099 | 16072 |
+| 2 | 535205 | 0.66 | 0.66 | 307.5 | 10043 | 27515 |
