+++ docs/paper.md
@@ -73,11 +73,14 @@ This package can be applied to any tree in which the nodes
 are labeled by positive integers. I.e., although it was designed and
 tested mainly for the NCBI taxonomy tree, it is not limited to it.
 
-It can be installed using ``pip install fastsubtrees``. Its
-functionality can be used through either executable scripts or an API, which
-can be employed from other Python programs.
-Manuals are provided for both
-kind of interfaces. A complete test suite based on _pytest_ is provided.
+It can be installed using ``pip install fastsubtrees`` (provided that the
+``mariadb`` database and the ``mariadb`` pip package is installed in the system).
+A preinstalled
+version of the package, with all dependencies, is provided as a Docker image.
+The package functionality can be used
+through either executable scripts or an API, which can be employed from other
+Python programs. Manuals are provided for both kind of interfaces. A complete
+test suite based on _pytest_ is provided.
 
 The first step when using _fastsubtrees_ is constructing its tree
 representation and storing it to file. Any source can be used as input (e.g. tabular files, or
@@ -101,26 +104,28 @@ dynamic, i.e. it is possible to add and remove subtrees.
 
 # Subtree extraction benchmarks
 
-Benchmarks were performed on a Linux server with Intel Xeon E7-4850 2.0 Ghz
-CPU and 1 Tb RAM. Running times were measured as an average of 3 runs using
+Benchmarks were performed on a MacBook Pro 2021 with Apple M1 Pro CPU
+and 32 Gb RAM. Running times were measured as an average of 3 runs using
 GNU time version 1.9 [@GNUtime].
 For the tests Python version 3.10.2 was used.
+The tested version of fastsubtrees was 1.3.
 
-The NCBI taxonomy tree used for the tests was downloaded on May 28, 2022
+The NCBI taxonomy tree used for the tests was downloaded on October 7, 2022
 from the NCBI FTP website [@NCBI:FTP].
 For downloading and keeping up-to-date a copy of the dump files using Python
 we developed the _ntmirror_ package (located in the directory ``ntmirror`` of
-the source code repository, version 1.1.0)
-[@NTMIRROR] installable using ``pip install ntmirror``.
-The tree contained 2949637 nodes as of May 28, 2022. The generation of the tree
+the source code repository, version 1.3.0)
+[@NTMIRROR] installable using ``pip install ntmirror`` (if the ``mariadb``
+package was previously installed).
+The tree contained 2447574 nodes. The generation of the tree
 representation of the NCBI taxonomy tree from the dump files
 using the _fastsubtrees-construct_ script required
-23 minutes and 35 seconds (average of 3 runs).
+9 minutes and 23 seconds (average of 3 runs).
 
 An alternative to the use of _fastsubtrees_ is to store the tree data in a SQL
 database and extract subtrees using hierarchical SQL queries. We implemented
 this solution in _ntmirror_: the dump data is loaded into a MariaDB
-database (version 10.6.8) and the script _ntmirror-extract-subtree_, based on
+database (version 10.6.10) and the script _ntmirror-extract-subtree_, based on
 SQLAlchemy allows to extract a subtree using SQL.
 
 To select subtrees of different sizes for the benchmarks, we started from the
@@ -132,37 +137,37 @@ node (TaxID 2), i.e. nodes 83333 (_Escherichia coli_ K12),
 The running time and memory usage of _fastsubtree_ are compared with those for
 hierarchical SQL queries in Table 1.
 
-| Subtree Id | Subtree Size | Avg. CPU Time SQL (s) | Avg. CPU Time Fastsubtrees (s) | Avg. Real Time SQL (s) | Average Real Time Fastsubtrees (s) | Memory Peak SQL (kb) | Memory Peak Fastsubtrees (kb) |
[…]
