+++ ntmirror/ntmirror/__init__.py
@@ -1,1 +1,5 @@
+# ntmirror library
+
 from .downloader import Downloader
+
+VERSION="1.3.1"
