### 2022-10-07 6e6b36a
better error message when Python modules cannot be loaded
