### 2022-10-07 1ebce18
script to facilitate adding attribute values after adding a subtree
