### 2022-10-07 a30e61e
update and improve documentation
mention the new CLI scripts to edit attribute values
