+++ docs/CLI.md
@@ -50,6 +50,11 @@ In the first example, the function in ``my_module.py`` is called as
 ``element_parent_ids("a", "b", "c")``, in the second as
 ``element_parent_ids("x", k1="v1", k2="v=2")``.
 
+Some modules implementing the described interface are provided
+under ``fastsubtrees/ids_modules``. In particular, ``ids_from_database.py``
+provides an interface for reading the tree data from a database table,
+and ``ids_from_tabular_file.py`` from a tabular input file.
+
 ## Modifying an existing tree representation
 
 ### Adding a subtree
@@ -67,9 +72,6 @@ arguments are passed to the script:
   To pass keyword arguments, using the syntax ``key=value``, use the option
   ``--keyargs``.
 
-If attribute have been defined, as described in the following section,
-the attribute files are automatically detected and modified too.
-
 ### Deleting a subtree
 
 The script `fastsubtrees-delete-subtree` is used to delete one or multiple
@@ -82,9 +84,6 @@ The following parameters are used for the script:
 - `nodeid`: ID of a leaf node to be deleted, or of an internal node, i.e.
    the subtree root of the subtree to be deleted
 - `tree`: File containing the tree
-
-If attribute have been defined, as described in the following section,
-the attribute files are automatically detected and modified too.
 
 ## Tree attributes
 
@@ -112,10 +111,51 @@ parameters along with an optional input parameter to run this script:
   To pass keyword arguments, the syntax **key=value** and the option
   **--keyargs** has to be used.
 
+A module using the described ``attrmod`` interface for adding attributes
+from tabular files is provided under ``fastsubtrees/ids_modules`` and
+can be used as follows:
+```
+fastsubtrees-attr-construct my.tree myattribute \
+   fastsubtrees/ids_modules/attr_from_tabular_file.py tabularfile.tsv
+```
+By default, the IDs are supposed to be in column 0, the attribute values in
+column 1 and the columns to be tab-separated; different values can be
+provided as keyword arguments, e.g.:
+```
+fastsubtrees-attr-construct my.tree myattribute \
+   fastsubtrees/ids_modules/attr_from_tabular_file.py tabularfile.tsv
+   --keyargs id_col=2 attr_col=10 separator=';'
+```
+
+### Attributes when editing a tree
+
+If attribute have been defined, as described in the following section,
+the attribute files are automatically detected and modified too,
+when adding or deleting a subtree.
+
+When adding a subtree, the attribute will initially have an empty value (None)
+for each of the additional nodes. In order to add attribute values for the new
+nodes, the script ``fastsubtrees-attr-add`` is employed. This must be called
+for each of the attributes for which values shall be added, and uses the same
+interface as the ``fastsubtrees-attr-construct`` script (see above).
+
+### Editing attribute values
+
+To add new values for an attribute, use the ``fastsubtrees-attr-add`` script.
+The script uses the same
+interface as the ``fastsubtrees-attr-construct`` script (see above).
+
+By default, the new values of the attributes for a node are added to the
+existing ones. If the existing ones shall be replaced by the new ones,
+use the option ``--replace``.
+
+To remove the values of an attribute for a list of given nodes,
+use the ``fastsubtrees-attr-delete`` script.
+
 ### Removing an attribute
 
-To remove an attribute, just delete the file generated by
-`fastsubtrees-attr-construct`.
+To remove an attribute, use the ``fastsubtrees-attr-delete`` script
+with the option ``--all``.
 
 ## Subtree queries
 
