### 2022-10-07 d7b1a8a
fix subtree sizes after insertion
