### 2022-10-07 9c8498e
raise error if trying to delete not existing node
