### 2022-10-07 1ce1cf0
renamed version to 1.4
(since mistakenly uploaded version name 1.4 to pip,
which can't be undone)
