+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-=== 1.3.1 ===
+=== 1.4 ===
 - add scripts to modify attribute values and adding values after add-subtree
 - improved the documentation
 - distribute with package also module for attribute construction from tab files
