### 2022-10-07 b46a5aa
fix benchmarks_attr.sh keyword argument call
