### 2022-10-07 ggonnella/fastsubtrees#4 @KonradHoeffner
I tried it with the docker image but it didn't work:

```
$ docker build . -t fastsubtrees
$ docker run --rm fastsubtrees pytest
rootdir: /
plugins: console-scripts-1.3.1, dash-2.0.0
collected 0 items / 1 error

==================================== ERRORS ====================================
________________________ ERROR collecting test session _________________________
usr/local/lib/python3.8/dist-packages/_pytest/runner.py:338: in from_call
    result: Optional[TResult] = func()
usr/local/lib/python3.8/dist-packages/_pytest/runner.py:369: in <lambda>
    call = CallInfo.from_call(lambda: list(collector.collect()), "collect")
usr/local/lib/python3.8/dist-packages/_pytest/main.py:711: in collect
    for direntry in visit(str(argpath), self._recurse):
usr/local/lib/python3.8/dist-packages/_pytest/pathlib.py:667: in visit
    yield from visit(entry.path, recurse)
usr/local/lib/python3.8/dist-packages/_pytest/pathlib.py:667: in visit
    yield from visit(entry.path, recurse)
usr/local/lib/python3.8/dist-packages/_pytest/pathlib.py:667: in visit
    yield from visit(entry.path, recurse)
usr/local/lib/python3.8/dist-packages/_pytest/pathlib.py:652: in visit
    for entry in os.scandir(path):
E   FileNotFoundError: [Errno 2] No such file or directory: '/dev/fd/11'
=========================== short test summary info ============================
ERROR  - FileNotFoundError: [Errno 2] No such file or directory: '/dev/fd/11'
!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!
=============================== 1 error in 0.13s ===============================
```
