### 2022-10-08 #4755 @ggonnella
I created a docker as suggested by the reviewer.
This allows to run tests, benchmarks and the example application without hassles.
I improved the documentation, including the part regarding installation.

I updated the benchmark results in the paper, using values obtained using the Docker
mentioned above. In the repository are now scripts, which allow to reproduce the
tables of the paper fully automatically.
