### 2022-10-10 ggonnella/fastsubtrees#4 @vinisalazar
I was able to run tests with the `make test` command and also by installing from source. However, it would be better to be able to run the tests from the repository root, instead of inside the `tests` directory.
