### 2022-10-10 #4755 @vinisalazar
Hi,

I believe that the problem targeted by the software is legitimate and that the applications are solid. So, I'm inclined to accept this paper. However, I agree that there are a number of issues that must be addressed.

I was able to successfully install from both source and using the Docker image, and to run the tests.

The Docker image proved to be the easiest method. Although it did take long to build, once the container was ready, all commands could be executed. Still, I think many users will not be familiar with Docker, so installation from pip or some other package manager is preferable. Here comes in the problem documented in ggonnella/fastsubtrees#3, of the MariaDB Server dependency. I was able to install MariaDB Server locally using `brew`, but if it's a core dependency of the package, it must be packaged along with it. I think the preferable alternative for packaging fastsubtrees would be conda, specifically the [Bioconda channel](https://anaconda.org/bioconda/). Another problem here is that mariadb-server itself does not have a conda distribution (except for [this one](https://anaconda.org/ostrokach/mariadb-server), which hasn't been updated in 6 years), so one would have to be made for it, or MariaDB would need to be compiled from source in the fastsubtrees recipe build script (harder to do). This would allow one to specify mariadb-server as a dependency and to install fastsubtrees without Docker in a single command.

I insist on this point of installation because I think it's the main barrier for potential users of the package. If that is fixed, it will significantly improve the package accessibility. Additionally, it would be nice to have an additional level of detail on how to install MariaDB Server, or even why it's required, so users find it easier to do it themselves, if they have to.

Here are some other observations which I think would help improve the package:
* Add repository title as header of README file.
* Add a repository description on the "About" section of the GitHub repository page.
* Docs should be rendered into some sort of page (such as a [GitHub Wiki](https://docs.github.com/en/communities/documenting-your-project-with-wikis) or ReadTheDocs).
* CLI subcommands (`fastsubtrees-construct`, `-query`, `-add-subtree` etc) should be made into subparsers of a single command `fastsubtrees`, which when called upon lists all available single commands and a one-line description.
* It would be nice to have some CLI examples in the repository README.
* The requirements file should explicitly list the requirements. Having it as is (only content is `-e .`, presumably for `pip install` or a setup command) is bad practice. Either remove this file or modify it to list the requirements.
* Tests with `pytest` should ideally be run from the repository root instead of the `tests/` directory (related to ggonnella/fastsubtrees#4).
* Test coverage seems very low. After running tests with [coverage.py](https://coverage.readthedocs.io/), there was only 6% coverage across the package code.
  * <img width="882" alt="Screen Shot 2022-10-10 at 2 04 43 pm" src="https://user-images.githubusercontent.com/17276653/194794021-73c383bc-4b7e-4f0c-b530-fc207be26d2c.png">
* Lastly, since there already is a test suite, it would be nice to have a GitHub Actions workflow to run the tests, and perhaps an additional action for things like linting/code style/etc -- could be done with [pre-commit](https://pre-commit.com/).

I don't think it's mandatory to address all of these immediately, but they would certainly contribute to the overall state of the paper. Please let me know if you have any questions.

Kind regards,
Vini
