### 2022-10-10 #4755 @ggonnella
Thank you very much for your review and constructive comments. I think we can address several of the minor points.
I have indeed some questions/comments, regarding the other:
- the main installation issue seems to be related to MariaDB; actually MariaDB is not really required for _fastsubtrees_ to work. It is only required for the benchmarks, since the library is compared to a database-based solution. My idea to handle this would then be to remove the dependency to the "mariadb" package. This would allow installation using Pip (and thus also Conda) and to run the package benchmarks, except comparative benchmarks, for which the database would be required - but for this the Docker image would be then the way to go.
- Actually tests cover all operations, and definitely at least most of the code under ``tree.py`` and ``attribute.py`` (which contain most of the code), for which the report says "0%".
