### 2022-10-10 ggonnella/fastsubtrees#4 @KonradHoeffner
Yes, the tests should be able to be executed with a simple and single Docker command.
Even when going into the Docker container, it did not work:

```
$ docker run --rm -it fastsubtrees bin/bash
root@4829c1e94d5b:/# cd fastsubtrees/docker
root@4829c1e94d5b:/fastsubtrees/docker# make test
if [ -z "$(docker ps -a | grep "fastsubtreesC")" ]; then \
	docker run -p 8050:8050 --detach --name "fastsubtreesC" "fastsubtrees"; \
else \
	docker start "fastsubtreesC"; \
fi
/bin/sh: 1: docker: not found
/bin/sh: 2: docker: not found
make: *** [Makefile:42: container] Error 127
```
