### 2022-10-10 ggonnella/fastsubtrees#4 @KonradHoeffner
Or do you mean executing `make test` outside of Docker? This does not work as well:

```
docker$ make test
if [ -z "$(docker ps -a | grep "fastsubtreesC")" ]; then \
	docker run -p 8050:8050 --detach --name "fastsubtreesC" "fastsubtrees"; \
else \
	docker start "fastsubtreesC"; \
fi
8831a072d81b9c4a705ad09faf3b20bd20e350e8f931315b425713d213c25c52
docker exec -d "fastsubtreesC" bash
sleep 5
docker exec -it "fastsubtreesC" bash \
	-c "mariadb -u root -prootpass < /fastsubtrees/docker/create_db.sql"
ERROR 2002 (HY000): Can't connect to local server through socket '/run/mysqld/mysqld.sock' (2)
make: *** [Makefile:60: create_test_db] Error 1
```
