### 2022-10-10 ggonnella/fastsubtrees#4 @ggonnella
@KonradHoeffner yes, I meant outside of Docker.

he problem seems to be that the database server is not started on your system, which happens on mine, when I do ``docker exec -d "fastsubtreesC" bash`` (since I based the image on a MariaDB image) and wait some seconds on my system. I do not know yet, how to fix this.
