### 2022-10-10 e3b78de
remove ntmirror from requirements
it is required for the comparative benchmarks
and for setting up the example application
but not for the main functionality of the fastsubtrees library
