### 2022-10-10 #4755 @vinisalazar
@ggonnella sounds good, if you could separate MariaDB from the overall installation, that would be great. I ran the `make benchmarks_all` on my laptop and it finished it without any problems, so I think that the Docker image is working well for the comparative benchmarks, for folks who are keen to reproduce that.

> * Actually tests cover all operations, and definitely at least most of the code under `tree.py` and `attribute.py` (which contain most of the code), for which the report says "0%".

If that is the case, do you think it would be feasible to run the testing suite plus a coverage report on a GitHub Actions workflow?

Best wishes,
Vini
