### 2022-10-10 ggonnella/fastsubtrees#4 @KonradHoeffner
It works now! I don't know what the reason for the problem, maybe it was due to timing and 5 seconds weren't enough (I'm using an i9-12900k though), but I tried it several times now and it worked even when stopping the Docker container beforehand.

```
docker$ make test
if [ -z "$(docker ps -a | grep "fastsubtreesC")" ]; then \
	docker run -p 8050:8050 --detach --name "fastsubtreesC" "fastsubtrees"; \
else \
	docker start "fastsubtreesC"; \
fi
fastsubtreesC
docker exec -d "fastsubtreesC" bash
sleep 5
docker exec -it "fastsubtreesC" bash \
	-c "mariadb -u root -prootpass < /fastsubtrees/docker/create_db.sql"
docker exec -it "fastsubtreesC" bash \
	-c "cd /fastsubtrees && make tests"
pytest
==================================================================== test session starts =====================================================================
rootdir: /fastsubtrees
plugins: console-scripts-1.3.1, dash-2.0.0
collected 35 items

ntmirror/tests/test_dbload_api.py ..                                                                                                                   [  5%]
ntmirror/tests/test_dbload_cli.py ..                                                                                                                   [ 11%]
ntmirror/tests/test_downloader_api.py ...                                                                                                              [ 20%]
ntmirror/tests/test_downloader_cli.py ...                                                                                                              [ 28%]
tests/test_add_subtree.py ......                                                                                                                       [ 45%]
tests/test_cli.py ...                                                                                                                                  [ 54%]
tests/test_construction.py .......                                                                                                                     [ 74%]
tests/test_delete_subtree.py ...                                                                                                                       [ 82%]
tests/test_file_io.py ...                                                                                                                              [ 91%]
tests/test_query.py ...                                                                                                                                [100%]

==================================================================== 35 passed in 10.47s =====================================================================
```
