+++ fastsubtrees/__init__.py
@@ -1,6 +1,6 @@
 # Fastsubtrees library
 
-VERSION="1.4"
+VERSION="1.5"
 
 # setup loguru by disabling it, as expected for libraries
 from loguru import logger
