### 2022-10-10 ggonnella/fastsubtrees#3 @ggonnella
This is now fixed, since the dependency to MariaDB was removed when installing using pip.
