### 2022-10-11 #4755 @vinisalazar
Hi @ggonnella,

I rebuilt the Docker image today after the updates. I still can successfully run the benchmarks from the Docker container, but I was having trouble with the `genome-attributes-viewer` app, it started, I ran a couple of queries, and then it crashed with this traceback:

```
root@d8798c592652:/fastsubtrees/docker# ./start-example-app
Dash is running on http://0.0.0.0:8050/

 * Serving Flask app 'start' (lazy loading)
 * Environment: production
   WARNING: This is a development server. Do not use it in a production deployment.
   Use a production WSGI server instead.
 * Debug mode: on
./start-example-app: line 15:   823 Killed                  ./start.py
root@d8798c592652:/fastsubtrees/docker# Traceback (most recent call last):
  File "/fastsubtrees/genomes-attributes-viewer/start.py", line 171, in <module>
    app.run_server(debug=True, host='0.0.0.0')
  File "/usr/local/lib/python3.10/dist-packages/dash/dash.py", line 2033, in run_server
    self.server.run(host=host, port=port, debug=debug, **flask_run_options)
  File "/usr/local/lib/python3.10/dist-packages/flask/app.py", line 920, in run
    run_simple(t.cast(str, host), port, self, **options)
  File "/usr/local/lib/python3.10/dist-packages/werkzeug/serving.py", line 1000, in run_simple
    _rwr(
  File "/usr/local/lib/python3.10/dist-packages/werkzeug/_reloader.py", line 418, in run_with_reloader
    ensure_echo_on()
  File "/usr/local/lib/python3.10/dist-packages/werkzeug/_reloader.py", line 398, in ensure_echo_on
    termios.tcsetattr(sys.stdin, termios.TCSANOW, attributes)
termios.error: (5, 'Input/output error')
```

I was also able to run some `fastsubtrees-query` commands, which got me thinking it would be nice to have other attributes of a taxonomy entry (e.g. Scientific Name, Lineage, etc) in the output. Maybe this could be achieved by using [taxonkit](https://bioinf.shenwei.me/taxonkit/) along with fastsubtrees.

I maintain my impression that `fastsubtrees` has legitimate and useful applications, but I think users will struggle to adopt it given the current installation procedure and command interface. I think that maybe more focus could be given on having users easily (*i.e.* automatically) index and build the NCBI Taxonomy Tree and run queries on it. The `genome-attributes-viewer` app should also be able of being deployed automatically with a single command after install. Ideally, upon installing `fastsubtrees` with conda, users should already have a prebuilt NCBI Taxonomy Tree and could start the Dash app right away.

 I also struggled with the output of the `fastsubtrees-attr-query` command, for example:

```
root@d8798c592652:/fastsubtrees# fastsubtrees-attr-query ncbi-taxonomy.tree GC_content 1129
2022-10-11 03:51:44 SUCCESS: Tree loaded from file "ncbi-taxonomy.tree"
2022-10-11 03:51:44 INFO: Subtree of node 1129 has size 1762
[None, None, None, None, None, [0.49187], None, None, None, None, [0.633354, 0.63345], [0.524473], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.563698], None, None, None, None, [0.600861], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.493706], None, None, None, [0.485047], None, None, None, None, None, None, None, None, None, None, None, None, [0.491264], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.604614], None, None, None, [0.597736], None, None, None, None, [0.603592], None, [0.644399], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.64059], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.541645], None, None, [0.602373], None, None, None, None, [0.584503], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.492615], [0.493446], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.613678], None, [0.590853], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.606186], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.607883], None, None, None, None, None, None, None, None, [0.58221], None, [0.599123], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.534223], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.406211], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.574981], None, None, None, None, [0.646371], [0.67057], [0.645688], [0.636933], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.534347], None, None, [0.549756], [0.583811], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.603577], None, [0.613608], [0.59087], [0.588178], [0.592676], [0.539637], [0.607604], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.49424], None, [0.581548], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.639157], None, None, [0.538882], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.679814], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.625627], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.66265], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.631121], None, None, None, [0.618893], None, None, None, None, None, None, [0.554843, 0.554375], [0.554356], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [0.55123], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None]
```
This output is confusing to me. I understand that these are GC content values, but how can I see which value belongs to which node? When I redirect this to a file, for example, I get this Python representation of a list, where I guess it would be preferable to have a stdout of one value per line.

Best,
Vini
