### 2022-10-11 #4755 @ggonnella
Hi @vinisalazar,

- example app: I will try to understand where is the problem, I have no idea about this error, at the moment
- query command: this can be a good idea for the NCBI taxonomy tree, but I would like to keep the software library generic, since I think it can be very useful for other trees as well, besides the NCBI taxonomy tree - at least I plan to use it for other applications as well in the future; thus I think this can be implemented as an additional wrapper script/application, which uses the library
- the same I can tell for creating the tree for NCBI taxonomy; I can and will write a wrapper script which does what you suggest automatically, but I don't want to change the basic idea, that the library can be used for other trees as well
- genome-attribute-viewer is an example application, to show the possible kind of application of the library; the library itself is the focus of this paper, not the example application; installing with pip/conda should install the library only, on my opinion; also because there is now already the Docker (required by the other reviewer) which is doing exactly that - creating the NCBI tree, installing/starting the example application
- I agree regarding the output of attr-query and will change the default output accordingly
