### 2022-10-11 ggonnella/fastsubtrees#8 @KonradHoeffner
constructing tree data and index parallelizable?
Constructing the tree data and index takes several minutes and only seems to use one CPU core, is it possible to parallelize that?

```
fastsubtrees$ docker exec fastsubtrees benchmarks
# Downloading the NCBI taxonomy dump...
# Building the fastsubtrees NCBI taxonomy tree...
2022-10-11 08:09:41 INFO: Constructing tree from NCBI taxonomy dump file ntdumpdir/nodes.dmp
2022-10-11 08:09:41 INFO: Constructing temporary parents table...
2022-10-11 08:09:41 INFO: Reading data from file "ntdumpdir/nodes.dmp" ...
2447844it [00:01, 1349487.11it/s]
2022-10-11 08:09:43 INFO: Constructing subtree sizes table...
2984676it [00:06, 451387.87it/s]
2022-10-11 08:09:50 INFO: Constructing tree data and index...
100%|██████████| 2984676/2984676 [05:57<00:00, 8351.77it/s]
2022-10-11 08:15:47 SUCCESS: Tree data structure constructed
2022-10-11 08:15:47 SUCCESS: Tree constructed
2022-10-11 08:15:47 SUCCESS: Tree written to file "ncbi-taxonomy.tree"
```
