### 2022-10-11 ggonnella/fastsubtrees#15 @ggonnella
tests from the repository root
@vinisalazar wrote:

> Tests with pytest should ideally be run from the repository root instead of the tests/ directory (related to https://github.com/ggonnella/fastsubtrees/issues/4).
