### 2022-10-11 ggonnella/fastsubtrees#15 @ggonnella
This actually works correctly on systems I can access (and it worked also in previous versions), i.e. writing "pytest" in the main repository directory runs the tests of fastsubtrees and those of ntmirror (by default skipping the database-dependant tests of the latter).
