### 2022-10-11 ggonnella/fastsubtrees#16 @ggonnella
test coverage
@vinisalazar wrote:
> Test coverage seems very low. After running tests with [coverage.py](https://coverage.readthedocs.io/), there was only 6% coverage across the package code.
