### 2022-10-11 ggonnella/fastsubtrees#16 @ggonnella
I suppose this is realated to the tests being implemented using pytest. I will look for this into more detail.
The solution is probably to use the _pytest-cov_ module to measure the coverage, and not _coverage.py_.
