### 2022-10-11 ggonnella/fastsubtrees#17 @ggonnella
Github actions
@vinisalazar wrote:

> it would be nice to have a GitHub Actions workflow to run the tests, and perhaps an additional action for things like linting/code style/etc -- could be done with [pre-commit](https://pre-commit.com/).
