### 2022-10-11 ggonnella/fastsubtrees#19 @ggonnella
fastsubtrees-query: scientific names / lineage in the output
@vinisalazar wrote:
> I was also able to run some fastsubtrees-query commands, which got me thinking it would be nice to have other attributes of a taxonomy entry (e.g. Scientific Name, Lineage, etc) in the output. Maybe this could be achieved by using [taxonkit](https://bioinf.shenwei.me/taxonkit/) along with fastsubtrees.
