### 2022-10-11 a9a79c6
implement CLI using subcommands
