+++ tests/test_cli.py
@@ -8,16 +8,16 @@ import os
 
 @pytest.mark.script_launch_mode('subprocess')
 def test_construct(testout, testdata, ids_modules, script, script_runner):
-  args = [testout("small_tree.tree"), ids_modules("ids_from_tabular_file.py"),
-          testdata("small_tree.tsv")]
-  ret = script_runner.run(script("fastsubtrees-construct"), *args)
+  args = ["construct", testout("small_tree.tree"),
+          ids_modules("ids_from_tabular_file.py"), testdata("small_tree.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert os.path.exists(testout("small_tree.tree"))
   # add an attribute
-  args = [testout("small_tree.tree"), 'attrX',
+  args = ["attr", "construct", testout("small_tree.tree"), 'attrX',
           ids_modules("attr_from_tabular_file.py"),
           testdata("small_tree_attrX.tsv")]
-  ret = script_runner.run(script("fastsubtrees-attr-construct"), *args)
+  ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert os.path.exists(testout("small_tree.tree.attrX.attr"))
 
@@ -26,22 +26,22 @@ def test_query(testout, testdata, ids_modules, script, script_runner,
                results_query_small_tree_id_8,
                results_query_small_tree_id_8_attrX):
   # construct the tree
-  args = [testout("small_tree.tree"), ids_modules("ids_from_tabular_file.py"),
-          testdata("small_tree.tsv")]
-  ret = script_runner.run(script("fastsubtrees-construct"), *args)
+  args = ["construct", testout("small_tree.tree"),
+          ids_modules("ids_from_tabular_file.py"), testdata("small_tree.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
   # add an attribute
-  args = [testout("small_tree.tree"), 'attrX',
+  args = ["attr", "construct", testout("small_tree.tree"), 'attrX',
           ids_modules("attr_from_tabular_file.py"),
           testdata("small_tree_attrX.tsv")]
-  ret = script_runner.run(script("fastsubtrees-attr-construct"), *args)
+  ret = script_runner.run(script("fastsubtrees"), *args)
   # query the constructed tree
-  args = [testout("small_tree.tree"), "8"]
-  ret = script_runner.run(script("fastsubtrees-query"), *args)
+  args = ["query", testout("small_tree.tree"), "8"]
+  ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout == "".join(f"{x}\n" for x in results_query_small_tree_id_8)
   # query the attribute
-  args = [testout("small_tree.tree"), "attrX", "8"]
-  ret = script_runner.run(script("fastsubtrees-attr-query"), *args)
+  args = ["attr", "query", testout("small_tree.tree"), "attrX", "8"]
+  ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout.strip() == str(results_query_small_tree_id_8_attrX)
 
@@ -53,65 +53,67 @@ def test_edit(testout, testdata, ids_modules, script, script_runner,
               results_query_small_tree_id_8_attrX_after_add,
               results_query_small_tree_id_8_attrX_after_add2):
   # construct the tree
-  args = [testout("small_tree.tree"), ids_modules("ids_from_tabular_file.py"),
-          testdata("small_tree.tsv")]
-  ret = script_runner.run(script("fastsubtrees-construct"), *args)
[…]
