+++ setup.py
@@ -34,14 +34,7 @@ setup(name='fastsubtrees',
         'Topic :: Software Development :: Libraries',
       ],
       packages=find_packages(),
-      scripts=['bin/fastsubtrees-construct',
-               'bin/fastsubtrees-query',
-               'bin/fastsubtrees-attr-construct',
-               'bin/fastsubtrees-attr-query',
-               'bin/fastsubtrees-attr-add',
-               'bin/fastsubtrees-attr-delete',
-               'bin/fastsubtrees-add-subtree',
-               'bin/fastsubtrees-delete-subtree'],
+      scripts=['bin/fastsubtrees'],
       zip_safe=False,
       include_package_data=True,
       install_requires=['tqdm>=4.57.0', 'loguru>=0.5.1', 'docopt>=0.6.2',
