### 2022-10-11 55e6e66
README.md, correction to first sentence
