### 2022-10-11 bdf3d0c
add a --tab option to fastsubtrees construct
this simplify use when the input is a TSV file
