### 2022-10-11 10f8e4f
add CLI examples to the README
