### 2022-10-11 522b981
option --tab to simplify use for tabular file input
