### 2022-10-11 7da98a0
fix the benchmark scripts after changes in the CLI
