+++ benchmarks/benchmarks_attr.sh
@@ -35,7 +35,7 @@ for attr in genome_size GC_content; do
   for ((i=0; i<$NREPEATS; i++)); do
     STEP="construct-$attr"
     /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-      fastsubtrees-attr-construct $TREE $attr \
+      fastsubtrees attr construct $TREE $attr \
         $FST_MOD_DIR/attr_from_tabular_file.py --datatype ${attr_dt[$attr]} \
           --keyargs \
             filename=$FST_DATA_DIR/accession_taxid_attribute.tsv.gz \
@@ -47,7 +47,7 @@ for attr in genome_size GC_content; do
     for ROOT in $NODES; do
       echo "Step $STEP from node $ROOT, iteration $i..."
       /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
-        fastsubtrees-attr-query --filter --countN --countV \
+        fastsubtrees attr query --countN --countV \
           $TREE $attr $ROOT > \
         $OUTDIR/attr_values.$attr.$ROOT
     done
