### 2022-10-11 1961e5e
better formatting of the attr query output
