### 2022-10-11 938cbe9
attr query, implement option --ids
