+++ fastsubtrees/commands/attr_query.py
@@ -12,6 +12,8 @@ Arguments:
 
 Options:
   --quiet      disable log messages
+  --ids        show node IDs alongside the attribute values
+               (output format: node_id <TAB> attribute_value)
   --nones      do not filter out None values in the output
   --counts     count the number of nodes with attributes and of attribute values
   --debug      print debug information
@@ -20,13 +22,6 @@ Options:
 """
 from docopt import docopt
 from fastsubtrees import Tree, logger, _scripts_support, VERSION, attribute
-
-def print_results(results):
-  for result in results:
-    if isinstance(result, list):
-      print(", ".join(str(r) for r in result))
-    else:
-      print(result)
 
 def main(args):
   logger.debug("Loading tree from file '{}'".format(args['<tree>']))
@@ -37,8 +32,7 @@ def main(args):
       format(args["<attribute>"], attributefile))
   attribute_list = attribute.get_attribute_list(tree, subtree_root, \
       attributefile)
-  filtered = None
-  if (not args["--nones"]) or args["--counts"]:
+  if args["--counts"]:
     filtered = [a for a in attribute_list if a is not None]
     if args["--counts"]:
       count = len(filtered)
@@ -46,10 +40,19 @@ def main(args):
       flattened = [e for sl in filtered for e in sl]
       count = len(flattened)
       logger.info("Number of attribute values: {}".format(count))
-  if args["--nones"]:
-    print_results(attribute_list)
-  else:
-    print_results(filtered)
+  node_ids = tree.subtree_ids(subtree_root) if args["--ids"] else None
+  for i, result in enumerate(attribute_list):
+    if result is None and not args["--nones"]:
+      continue
+    if isinstance(result, list):
+      attrstr = ", ".join(str(r) for r in result)
+    else:
+      attrstr = str(result)
+    if node_ids:
+      node_id = node_ids[i]
+      print("{}\t{}".format(node_id, attrstr))
+    else:
+      print(attrstr)
 
 if __name__ == "__main__":
   args = docopt(__doc__, version=VERSION)
