### 2022-10-11 9d11a90
document the --ids and --nones option of attr query
