+++ README.md
@@ -170,7 +170,7 @@ Once the attributes are created, their values in any subtree can be
 easily queried:
 ```
 # query the genome size values under node 562
-fastsubtrees attr query nt.tree genome_size 562
+fastsubtrees attr query --ids nt.tree genome_size 562
 ```
 
 #### CLI example with generic data
