### 2022-10-11 1a6bb08
correctly test output of attr query CLI tool
