### 2022-10-11 ggonnella/fastsubtrees#3 @ggonnella
The reason is that I didn't upload the newer version correctly to Pypi.

I created a version 1.6 from the current stand and uploaded it, so now it will work.
A cache purge is not needed.
