### 2022-10-11 ggonnella/fastsubtrees#3 @KonradHoeffner
Thanks, that works for me! However it did require a cache purge.

```
~$ pip install fastsubtrees
Defaulting to user installation because normal site-packages is not writeable
Collecting fastsubtrees
  Using cached fastsubtrees-1.4-py3-none-any.whl (21 kB)
Collecting tqdm>=4.57.0
  Using cached tqdm-4.64.1-py2.py3-none-any.whl (78 kB)
Collecting loguru>=0.5.1
  Using cached loguru-0.6.0-py3-none-any.whl (58 kB)
Collecting docopt>=0.6.2
  Using cached docopt-0.6.2.tar.gz (25 kB)
  Preparing metadata (setup.py) ... done
Collecting sh>=1.14.2
  Using cached sh-1.14.3.tar.gz (62 kB)
  Preparing metadata (setup.py) ... done
Requirement already satisfied: PyYAML>=6.0 in /usr/lib/python3.10/site-packages (from fastsubtrees) (6.0)
Collecting ntmirror>=1.2
  Using cached ntmirror-1.3.0-py3-none-any.whl (11 kB)
Collecting schema>=0.7.4
  Using cached schema-0.7.5-py2.py3-none-any.whl (17 kB)
Collecting sqlalchemy-repr
  Using cached sqlalchemy_repr-0.0.2-py3-none-any.whl (3.8 kB)
Collecting snacli
  Downloading snacli-1.2-py3-none-any.whl (4.7 kB)
Collecting sqlalchemy
  Using cached SQLAlchemy-1.4.41-cp310-cp310-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl (1.6 MB)
Collecting mysql
  Downloading mysql-0.0.3-py3-none-any.whl (1.2 kB)
Collecting mariadb
  Using cached mariadb-1.1.4.zip (97 kB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [19 lines of output]
      /bin/sh: line 1: mariadb_config: command not found
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/tmp/pip-install-m06k1h2u/mariadb_6cebeda013ed48baa93317c0b0d15caf/setup.py", line 27, in <module>
          cfg = get_config(options)
        File "/tmp/pip-install-m06k1h2u/mariadb_6cebeda013ed48baa93317c0b0d15caf/mariadb_posix.py", line 62, in get_config
          cc_version = mariadb_config(config_prg, "cc_version")
        File "/tmp/pip-install-m06k1h2u/mariadb_6cebeda013ed48baa93317c0b0d15caf/mariadb_posix.py", line 28, in mariadb_config
          raise EnvironmentError(
      OSError: mariadb_config not found.

      This error typically indicates that MariaDB Connector/C, a dependency which
      must be preinstalled, is not found.
      If MariaDB Connector/C is not installed, see installation instructions
      If MariaDB Connector/C is installed, either set the environment variable
      MARIADB_CONFIG or edit the configuration file 'site.cfg' to set the
       'mariadb_config' option to the file location of the mariadb_config utility.

      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
~$ pip cache purge && pip install fastsubtrees
Files removed: 24
Defaulting to user installation because normal site-packages is not writeable
Collecting fastsubtrees
  Downloading fastsubtrees-1.6-py3-none-any.whl (13 kB)
Requirement already satisfied: PyYAML>=6.0 in /usr/lib/python3.10/site-packages (from fastsubtrees) (6.0)
Collecting sh>=1.14.2
  Downloading sh-1.14.3.tar.gz (62 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 62.9/62.9 kB 2.1 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... done
Collecting schema>=0.7.4
  Downloading schema-0.7.5-py2.py3-none-any.whl (17 kB)
Collecting loguru>=0.5.1
  Downloading loguru-0.6.0-py3-none-any.whl (58 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 58.3/58.3 kB 4.2 MB/s eta 0:00:00
Collecting docopt>=0.6.2
  Downloading docopt-0.6.2.tar.gz (25 kB)
  Preparing metadata (setup.py) ... done
Collecting tqdm>=4.57.0
  Downloading tqdm-4.64.1-py2.py3-none-any.whl (78 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 78.5/78.5 kB 5.3 MB/s eta 0:00:00
Collecting contextlib2>=0.5.5
  Downloading contextlib2-21.6.0-py2.py3-none-any.whl (13 kB)
Using legacy 'setup.py install' for docopt, since package 'wheel' is not installed.
Using legacy 'setup.py install' for sh, since package 'wheel' is not installed.
Installing collected packages: sh, docopt, tqdm, loguru, contextlib2, schema, fastsubtrees
  Running setup.py install for sh ... done
  Running setup.py install for docopt ... done
Successfully installed contextlib2-21.6.0 docopt-0.6.2 fastsubtrees-1.6 loguru-0.6.0 schema-0.7.5 sh-1.14.3 tqdm-4.64.1
```
