### 2022-10-11 d612912
separate ntdownload and ntmirror
