+++ ntmirror/ntmirror/__init__.py
@@ -1,5 +1,3 @@
 # ntmirror library
 
-from .downloader import Downloader
-
-VERSION="1.3.1"
+VERSION="1.4"
