+++ ntdownload/ntdownload/__init__.py
@@ -1,0 +1,5 @@
+# ntdownload library
+
+from .downloader import Downloader
+
+VERSION="1.4"
