+++ ntdownload/tests/conftest.py
@@ -1,0 +1,15 @@
+#
+# (c) 2022 Giorgio Gonnella, University of Goettingen, Germany
+#
+import pytest
+import os
+from pathlib import Path
+
+@pytest.fixture
+def testout():
+  return lambda fn: Path(__file__).parent / 'output' / fn
+
+@pytest.fixture
+def script():
+  return lambda fn: Path(__file__).parent.parent / 'bin' / fn
+
