+++ ntdownload/setup.py
@@ -11,15 +11,14 @@ import sys
 if not sys.version_info[0] == 3:
   sys.exit("Sorry, only Python 3 is supported")
 
-setup(name='ntmirror',
+setup(name='ntdownload',
       version='1.4',
-      description="Easily updatable local NCBI taxonomy database copy",
+      description="Easily updatable local NCBI taxonomy dumps file copy",
       long_description=readme(),
       long_description_content_type="text/markdown",
-      install_requires=["sqlalchemy", "sqlalchemy_repr", "mysql", "mariadb",
-        "loguru", "snacli", "schema", "ntdownload"],
-      url='https://github.com/ggonnella/fastsubtrees/tree/main/ntmirror',
-      keywords="ncbi taxonomy database mirror",
+      install_requires=["loguru", "snacli", "schema"],
+      url='https://github.com/ggonnella/fastsubtrees/tree/main/ntdowload',
+      keywords="ncbi taxonomy dumpfiles mirror",
       author='Giorgio Gonnella',
       author_email='gonnella@zbh.uni-hamburg.de',
       license='ISC',
@@ -37,8 +36,7 @@ setup(name='ntmirror',
         'Topic :: Software Development :: Libraries',
       ],
       packages=find_packages(),
-      scripts=['bin/ntmirror-dbload',
-               'bin/ntmirror-extract-subtree'],
+      scripts=['bin/ntdownload'],
       zip_safe=False,
       test_suite="pytest",
       include_package_data=True,
