+++ ntdownload/CHANGELOG.txt
@@ -1,0 +1,3 @@
+== 1.4 ==
+
+- splitted from NtMirror
