### 2022-10-11 444a30b
update paper text to reflect changes
- versions
- separate ntdownload/ntmirror
- only installation of ntmirror requires MariaDB
- new CLI syntax
