+++ ntdownload/requirements.txt
@@ -1,1 +1,3 @@
--e .
+loguru
+snacli
+schema
