+++ requirements.txt
@@ -1,1 +1,6 @@
--e .
+tqdm>=4.57.0
+loguru>=0.5.1
+docopt>=0.6.2
+schema>=0.7.4
+sh>=1.14.2
+PyYAML>=6.0
