### 2022-10-11 3d7a918
transform Genomes Attributes Viewer into a package
