### 2022-10-11 05f9984
fix Docker image construction and create-tree-file script
