### 2022-10-11 ggonnella/fastsubtrees#16 @ggonnella
installing _pytest-cov_ by ``pip install pytest-cov`` and running
``pytest --cov=fastsubtrees -v tests/``
(I put both into a new Makefile goal, test-cov) returns the following:
```
Name                                                 Stmts   Miss  Cover
fastsubtrees/__init__.py                                16      5    69%
fastsubtrees/_scripts_support.py                        50     50     0%
fastsubtrees/attribute.py                               37     28    24%
fastsubtrees/commands/__init__.py                        0      0   100%
fastsubtrees/commands/add.py                            25     25     0%
fastsubtrees/commands/attr_add.py                       41     41     0%
fastsubtrees/commands/attr_construct.py                 31     31     0%
fastsubtrees/commands/attr_delete.py                    26     26     0%
fastsubtrees/commands/attr_query.py                     33     33     0%
fastsubtrees/commands/construct.py                      31     31     0%
fastsubtrees/commands/delete.py                         16     16     0%
fastsubtrees/commands/query.py                          13     13     0%
fastsubtrees/error.py                                   10      0   100%
fastsubtrees/ids_modules/__init__.py                     0      0   100%
fastsubtrees/ids_modules/attr_from_tabular_file.py      16      1    94%
fastsubtrees/ids_modules/ids_from_database.py           12     12     0%
fastsubtrees/ids_modules/ids_from_tabular_file.py        9      1    89%
fastsubtrees/tree.py                                   222     31    86%
TOTAL                                                  588    344    41%
```

It must be considered that the ``commands`` modules are all tested, using the ``pytest-console-script``, since they are subcommands of the ``fastsubtrees`` script; that's probably not catched up by pytest-cov. For the library main files ``tree.py`` and ``attribute.py`` the reported coverage is, respectively 86% and 24%. However, this again does not take account of the CLI tests, which actually check the entire functionality of these modules.
