+++ ntmirror/ntmirror/__init__.py
@@ -1,3 +1,3 @@
 # ntmirror library
 
-VERSION="1.4"
+VERSION="1.4.1"
