+++ genomes_attributes_viewer/AUTHORS.txt
@@ -1,0 +1,5 @@
+Giorgio Gonnella (project founder, developer)
+Aman Modi (developer)
+
+# Note: this file contains the list of the main developers of the project.
+# Minor contributors, patches, etc, are added to CONTRIBUTORS.txt.
