### 2022-10-11 037b796
update README genomes attributes viewer section
