### 2022-10-11 #4755 @ggonnella
@vinisalazar
I addressed your concerns regarding the complexity of installation. Now all it takes to run the example application is to install it by a single command (``pip install genomes_attributes_viewer``) and run it by a single command (``genomes-attributes-viewer``) which automatically does the setup steps the first time it is called.
