+++ docs/paper.bib
@@ -27,11 +27,20 @@
 
 @misc{NTMIRROR,
   author = {Giorgio Gonnella},
-  title = {Ntmirror Python package},
+  title = {NtMirror Python package},
   year = {2022},
   publisher = {Github},
   journal = {Github repository},
   url = {https://github.com/ggonnella/fastsubtrees/tree/main/ntmirror}
+}
+
+@misc{NTDOWNLOAD,
+  author = {Giorgio Gonnella},
+  title = {NtDownload Python package},
+  year = {2022},
+  publisher = {Github},
+  journal = {Github repository},
+  url = {https://github.com/ggonnella/fastsubtrees/tree/main/ntdownload}
 }
 
 @misc{FastsubtreesRepo,
