### 2022-10-11 bbbe342
allow multiple attributes query
