+++ ntdownload/ntdownload/__init__.py
@@ -2,4 +2,4 @@
 
 from .downloader import Downloader
 
-VERSION="1.4"
+VERSION="1.5"
