+++ ntdownload/CHANGELOG.txt
@@ -1,3 +1,5 @@
-== 1.4 ==
+== 1.5 ==
+- ntnames script to extract scientific names from the dumps
 
+== 1.4 ==
 - splitted from NtMirror
