### 2022-10-11 #4755 @ggonnella
@vinisalazar
I added functionality to add taxonomic names to the output, without breaking the generality of the software. This is done by adding the names as an attribute to the tree. An example of creating the names and showing them in a query is given in README.
