### 2022-10-11 #4755 @vinisalazar
Hi @ggonnella thank you for the updates.

> * query command: this can be a good idea for the NCBI taxonomy tree, but I would like to keep the software library generic, since I think it can be very useful for other trees as well, besides the NCBI taxonomy tree - at least I plan to use it for other applications as well in the future; thus I think this can be implemented as an additional wrapper script/application, which uses the library
> * the same I can tell for creating the tree for NCBI taxonomy; I can and will write a wrapper script which does what you suggest automatically, but I don't want to change the basic idea, that the library can be used for other trees as well

That is perfectly fine. I don't think it's a necessity to have this greater focus on the NCBI Taxonomy, since the underlying goal of the library is a different one. But, from a user's perspective, it would be nice to have easy access to that tree (i.e. having it be built automatically or even packaged along with the software).

I will wait for ggonnella/fastsubtrees#16, ggonnella/fastsubtrees#17 and ggonnella/fastsubtrees#19 to be finished and then I should be able to finish my review.

Best,
V
