### 2022-10-12 ggonnella/fastsubtrees#19 @ggonnella
Solved by 689d29e
