### 2022-10-12 ggonnella/fastsubtrees#19 @ggonnella
see example in README.md
