### 2022-10-12 b1064e3
add CITATION.cff, currently citing the repository
this will change if/when the paper is accepted
