+++ CITATION.cff
@@ -1,0 +1,10 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Gonnella"
+  given-names: "Giorgio"
+  orcid: "https://orcid.org/0000-0003-3900-5397"
+title: 'Fastsubtrees: simple and efficient subtrees extractions in Python with applications to NCBI taxonomy'
+version: 1.7
+date-released: 2022-10-12
+url: "htts://github.com/ggonnella/fastsubtrees/"
