+++ ntmirror/setup.py
@@ -17,7 +17,7 @@ setup(name='ntmirror',
       long_description=readme(),
       long_description_content_type="text/markdown",
       install_requires=["sqlalchemy", "sqlalchemy_repr", "mysql", "mariadb",
-        "loguru", "snacli", "schema", "ntdownload"],
+        "loguru", "snacli", "schema", "ntdownload", "PyYAML"],
       url='https://github.com/ggonnella/fastsubtrees/tree/main/ntmirror',
       keywords="ncbi taxonomy database mirror",
       author='Giorgio Gonnella',
