+++ requirements.txt
@@ -3,4 +3,3 @@ loguru>=0.5.1
 docopt>=0.6.2
 schema>=0.7.4
 sh>=1.14.2
-PyYAML>=6.0
