+++ ntmirror/requirements.txt
@@ -6,3 +6,4 @@ loguru
 snacli
 schema
 ntdownload
+PyYAML
