### 2022-10-12 b197c46
improve logger messages/levels in tree.py
