### 2022-10-12 c4f17df
unify query and attr query in a single command
