+++ tests/test_cli.py
@@ -43,12 +43,12 @@ def test_query(testout, testdata, ids_modules, script, script_runner,
           testdata("small_tree_attrX.tsv")]
   ret = script_runner.run(script("fastsubtrees"), *args)
   # query the constructed tree
-  args = ["query", testout("small_tree.tree"), "8"]
+  args = ["query", testout("small_tree.tree"), "8", "-N"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout == "".join(f"{x}\n" for x in results_query_small_tree_id_8)
   # query the attribute
-  args = ["attr", "query", testout("small_tree.tree"), "attrX", "8"]
+  args = ["query", testout("small_tree.tree"), "8", "attrX", "-a", "-N"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout.strip() == str(results_query_small_tree_id_8_attrX)
@@ -75,7 +75,7 @@ def test_edit(testout, testdata, ids_modules, script, script_runner,
           testdata("small_tree_add_subtree.tsv")]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
-  args = ["query", testout("small_tree.tree"), "8"]
+  args = ["query", testout("small_tree.tree"), "8", "-N"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout == "".join(f"{x}\n" for x in
@@ -86,13 +86,13 @@ def test_edit(testout, testdata, ids_modules, script, script_runner,
           testdata("small_tree_add_subtree2.tsv")]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
-  args = ["query", testout("small_tree.tree"), "8"]
+  args = ["query", testout("small_tree.tree"), "8", "-N"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout == "".join(f"{x}\n" for x in
       results_query_small_tree_id_8_add_subtree2)
   # attribute query before changing it
-  args = ["attr", "query", "--nones", testout("small_tree.tree"), "attrX", "8"]
+  args = ["query", testout("small_tree.tree"), "8", "attrX", "-a", "-N", "-n"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout.strip() == \
@@ -104,7 +104,7 @@ def test_edit(testout, testdata, ids_modules, script, script_runner,
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   # attribute query after changing it
-  args = ["attr", "query", "--nones", testout("small_tree.tree"), "attrX", "8"]
+  args = ["query", testout("small_tree.tree"), "8", "attrX", "-a", "-N", "-n"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout.strip() == \
@@ -113,14 +113,14 @@ def test_edit(testout, testdata, ids_modules, script, script_runner,
   args = ["delete", testout("small_tree.tree"), "10"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
-  args = ["query", testout("small_tree.tree"), "8"]
+  args = ["query", testout("small_tree.tree"), "8", "-N"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
[…]
