### 2022-10-12 c4e7fbb
query: keyword for using the tree root as subtree root
