### 2022-10-12 d4965d1
implement --subtree-sizes option for query
