### 2022-10-12 19045ed
correctly handle attribute creation after node deletion
