+++ CHANGELOG.txt
@@ -2,6 +2,7 @@
 - multiple attributes can be queried at once
 - explain how to load taxa names as attributes in README
 - simplification of the CLI: unify 'attr query' and 'query' subcommands
+- implement tree update and unify 'add' and 'delete' with it in the CLI
 
 === 1.7 ===
 - genomes_attributes_viewer is now a Python package, installable by pip

[…]
