### 2022-10-12 0fee07f
unify attribute creation/modification/deletion commands
