+++ fastsubtrees/commands/attribute.py
@@ -1,0 +1,133 @@
+#!/usr/bin/env python3
+"""
+Manage attribute values for trees.
+
+creation mode (default):
+  Given a source of attribute values, it will create a new attribute file
+  for the given tree, replacing any existing file.
+
+add mode (--add option):
+  Given a source of attribute values, it will add further values to the
+  attribute file for the given tree. If a node already has values for
+  the attribute, the new values will be appended to the existing one.
+
+replace mode (--replace option):
+  Given a source of attribute values, it will set values to the
+  attribute file for the given tree. If a node already has values for
+  the attribute, the new values will replace the existing one.
+
+delete mode (--delete option):
+  Delete the attribute file for the given tree, or, if one or more
+  node IDs are given, delete the attribute values for those nodes.
+
+Usage:
+  fastsubtrees attribute [options] <tree> <attribute> --tab <tabfile> [<attrmod_data>...]
+  fastsubtrees attribute [options] <tree> <attribute> <attrmod> [<attrmod_data>...]
+  fastsubtrees attribute [options] <tree> <attribute> --delete [<node_id>...]
+
+Using a tabular file as source of attribute values:
+
+  To use a tabular file <tabfile> as source of IDs, use the --tab option.
+  The file shall contain lines in the form "node_id <TAB> parent_id".
+
+  Optional arguments after <tabfile> are given in this order
+  or passed as key=value:
+    - id_col: column index of the element ID (default: 0)
+    - attr_col: column index of the attribute value (default: 1)
+    - separator: separator (default: tab)
+    - comment_pfx: prefix of lines to be skipped (default: '#')
+
+Using an attribute values source module:
+
+  The <attrmod> argument is the path to a Python module.
+  The module must have a function named attribute_values which yields
+  tuples (node_id, attribute_value).
+
+  The <attrmod_data> arguments are passed to the attribute_values function.
+  If they contain a = sign, they are interpreted as keyword arguments.
+
+Options:
+  --replace      replace attribute values in an existing file
+  --add          add new attribute values to an existing file
+  --strict       exit with an error if a node ID is not found in the tree
+  --delete       delete attribute values from an existing file
+  --tab          use a tabular file as source of input data;
+                 shorthand for using the module
+                 fastsubtrees/ids_modules/attr_from_tabular_file.py
+  --quiet        disable log messages
+  --debug        print debug information
+  --help         show this help message and exit
+  --version      show program's version number and exit
+  --datatype DT  function to apply to the attribute values before writing
+                 them to the output file; <DT> must be a valid Python function;
+                 the function can optionally be defined in the module specified
+                 by <attrmod> [default: str]
+"""
+
+from docopt import docopt
+from fastsubtrees import _scripts_support, logger, Tree, attribute, VERSION
+from collections import defaultdict
+from pathlib import Path
+
+def read_attribute_values(m, posargs, keyargs, cast):
+  result = defaultdict(list)
+  for k, v in m.attribute_values(*posargs, **keyargs):
+    result[int(k)].append(cast(v))
+  return result
+
+def write_attribute_file(args, attrvalues):
+  logger.debug("Loading tree from file '{}'".format(args['<tree>']))
+  tree = Tree.from_file(args["<tree>"])
+  outfname = attribute.attrfilename(args["<tree>"], args["<attribute>"])
+  if args["--add"] or args["--replace"] or args["--delete"]:
+    existing = attribute.read_attribute_values(tree, outfname)
+    if args["--add"] or args["--replace"]:
+      replaced = set()
[…]
