### 2022-10-12 68f66a0
document CLI changes
