+++ docs/CLI.md
@@ -8,21 +8,12 @@ The list of subcommands is displayed using ``fastsubtrees --help``.
 Using ``--help`` after a subcommand (e.g. ``fastsubtrees construct --help``
 displays the syntax and details of the subcommand.
 
-There are two sets of subcommands. For working with the tree itself,
-the corresponding verb (construct, query, add, delete) is used as subcommand
-directly. For working with attributes (see below), the subcommand is
-``attr`` and the verb (construct, add, delete) is used as sub-subcommand,
-after attr (e.g. ``fastsubtrees attr construct``).
-
 The following subcommands are available:
 ```
   construct         Construct the tree data structure and save it to file
-  query             List IDs/attribute values of the subtree under a given node
-  add               Add new nodes to an existing tree data structure
-  delete            Remove nodes from an existing tree data structure
-  attr construct    Save values of an attribute to (some of the) nodes of a tree
-  attr add          Add additional attribute values for more nodes of a tree
-  attr delete       Delete some or all values of an attribute
+  update            Update an existing tree data structure
+  attribute         Create, update or delete node attributes
+  query             List node IDs and/or attributes in the subtree under a node
 ```
 
 ## Tree construction
@@ -72,15 +63,14 @@ from a tabular file are given in the modules in the
 The ``element_parent_ids()`` function usually needs parameters. These
 can be passed either as positional or as keyword parameters. To pass them
 as positional parameters, they are just used as further arguments,
-after the module name. To define them as keyword parameters, the
-``--keyargs`` option is used, and all further arguments in the form
-``key=value`` (i.e. where at least one = is present), are parsed and
-passed as keywords parameters.
+after the module name.
+To define them as keyword parameters, the
+use the form ``key=value``.
 
 Examples:
 ```
 fastsubtrees construct my.tree my_module.py a b c
-fastsubtrees construct my.tree my_module.py --keyargs k1=v1 k2=v=2 x
+fastsubtrees construct my.tree my_module.py k1=v1 k2=v=2 x
 ```
 
 In the first example, the function in ``my_module.py`` is called as
@@ -94,33 +84,35 @@ and ``ids_from_tabular_file.py`` from a tabular input file.
 
 ## Modifying an existing tree representation
 
-### Adding a subtree
+### Updating a tree
 
-The script `fastsubtrees add` adds one or multiple nodes to an already
-existing tree. The interface is identical to the generic tree construction
-interface of ``fastsubtrees construct`` (see above), i.e. the following
-arguments are passed to the script:
-- `tree`: File containing the tree that has to be updated.
-- `idsmod`: A Python module that defines the function ``element_parent_ids()``,
-   which optionally takes arguments (``<idsmod_data>``) and which yields pairs
-   of IDs for all nodes ``(element_id, parent_id)`` of the subtree to be added.
-- `idsmod_data` (optional): list of arguments to be passed to
-  ``element_parent_ids()``. By default they are passed as positional arguments.
-  To pass keyword arguments, using the syntax ``key=value``, use the option
-  ``--keyargs``.
+It is possible to modify a tree, by updating the representation, given
+a source of node and parent IDs, similar to the one for the tree creation.
+For this the command ``fastsubtrees update`` is used.
+Any node not yielded in the source will be deleted, and new nodes
+will be added.
 
-### Deleting a subtree
+### Adding nodes
 
-The script `fastsubtrees delete` is used to delete one or multiple
-nodes from an exiting tree representation. If the specified node is a leaf
-node, then only that node it is deleted. If it is an internal node, then the
-entire subtree is also deleted, i.e. the set of all the descendants of the
-specified node.
[…]
