### 2022-10-13 3d7411b
make add/delete/update tree more flexible
- moving nodes/subtrees to another parent is now allowed
- parents can be inserted after the children
- update can now accept nodes/subtree movements
- more efficient way to edit attribute files
