+++ fastsubtrees/tree.py
@@ -5,6 +5,7 @@ Computation of the information necessary for the subtree query
 import struct
 import array
 import sys
+from collections import defaultdict
 from fastsubtrees import logger, tqdm, error
 
 class Tree():
@@ -196,100 +197,140 @@ class Tree():
         new_subtree_ids.append(data)
     return new_subtree_ids
 
-  def add_subtree(self, generator, attributefilenames=[], skip_existing=False,
-                  rm_existing_set=None, list_added=None):
+  def add_nodes(self, generator, attrfilenames=[], skip_existing=False,
+                  rm_existing_set=None, list_added=None, total=None,
+                  edit_script=None):
     n_added = 0
-    for node_number, parent in generator:
+    pending = defaultdict(list)
+    logger.info("Tree root: " + str(self.root_id))
+    if edit_script is None:
+      edit_script = []
+    for node_number, parent in tqdm(generator, total=total):
+      if rm_existing_set is not None:
+        rm_existing_set.discard(node_number)
       if node_number <= 0:
         raise error.ConstructionError(\
             f"The node IDs must be > 0, found: {node_number}")
       elif parent <= 0:
         raise error.ConstructionError(\
-            f"The node IDs must be > 0, found: {parent}")
+            f"The node parent IDs must be > 0, found: {parent}")
       if node_number < len(self.parents):
         if self.treedata[self.coords[node_number]] == Tree.DELETED:
-          raise error.DeletedNodeError(\
-              f'Node {node_number} was already deleted once. ' + \
-              'Cannot add the same node again')
-        else:
-          if node_number == self.root_id or \
-              self.parents[node_number] != Tree.UNDEF:
-            if skip_existing:
-              if rm_existing_set is not None:
-                rm_existing_set.remove(node_number)
-              continue
             raise error.ConstructionError(\
-              f"Node {node_number} had already been added with parent " + \
-              f"{self.parents[node_number]}, cannot add it again with " + \
-              f"parent {parent}")
-      n_added += 1
-      if list_added is not None:
-        list_added.append(node_number)
-      inspos = self.__prepare_node_insertion(node_number, parent)
-      if attributefilenames:
-        self.__insert_none_in_attribute_list(inspos, attributefilenames)
-      self.__insert_node(node_number, inspos, parent)
-      self.__update_subtree_sizes(node_number)
+              f'Node {node_number} / parent {parent} was previously '+\
+              f'deleted and had a different parent '+\
+              f'({self.parents[node_number]})')
+        if node_number == self.root_id:
+          if skip_existing:
+            if node_number != parent:
+              raise error.ConstructionError(\
+                  f'Node {node_number} / parent {parent} already exists '+\
+                  f'as the root node')
+            continue
+          else:
+            raise error.ConstructionError(\
+                f"The root node {node_number} already exists")
+        if self.parents[node_number] != Tree.UNDEF:
+          if skip_existing:
+            if self.parents[node_number] != parent:
+              if parent >= len(self.parents) or self.coords[parent] == 0:
+                pending[parent].append(('move', node_number))
+              else:
+                self.__move_subtree(node_number, parent, edit_script)
+            continue
+          else:
+            if self.parents[node_number] != parent:
+              raise error.ConstructionError(\
+                f"Node {node_number} / parent {parent} already exists "+\
[…]
