### 2022-10-13 d44e0e0
improve err msg
