### 2022-10-13 cfa9c82
add package to automatically download/create NCBI taxonomy tree
