+++ ntsubtree/ntsubtree/__init__.py
@@ -1,0 +1,5 @@
+# NtSubtree package
+
+from .constants import __version__
+from .ntsubtree import __auto_init
+__auto_init()
