+++ ntsubtree/AUTHORS.txt
@@ -1,0 +1,4 @@
+Giorgio Gonnella (conceptualization, project supervision, software development)
+
+# Note: this file contains the list of the main developers of the project.
+# Minor contributors, patches, etc, are added to CONTRIBUTORS.txt.
