+++ ntsubtree/README.md
@@ -1,0 +1,3 @@
+NtSubtree is a package based on Fastsubtrees which automatically
+downloads and installs the NCBI taxonomy tree during setup.
+
