### 2022-10-13 478612e
fix attribute update and allow readding deleted nodes
