### 2022-10-14 8c99575
further unify and simplify CLI
