+++ fastsubtrees/commands/tree.py
@@ -1,0 +1,161 @@
+#!/usr/bin/env python3
+"""
+Create or edit a tree representation
+
+Usage:
+  fastsubtrees tree <treefile> [--new|--update|--add] <tabfile> [options]
+  fastsubtrees tree <treefile> [--new|--update|--add] --module M [<args>...] [options]
+  fastsubtrees tree <treefile> --delete <subtree_root>... [options]
+
+Actions:
+  --new      (default) create a new tree from the given IDs source
+  --update   update an existing tree to reflect the given IDs source;
+             the root cannot be changed
+  --add      add new nodes to an existing tree
+  --delete   remove leaves or subtrees from an existing tree
+
+Tabular file input:
+  <tabfile>        tabular file containing IDs of the nodes and their parents
+  --ncbi           the tabular file is in NCBI taxonomy dump format
+  --separator S    separator in the tabular file (default: tab)
+  --elementscol E  column with the IDs of the elements (1-based, default: 1)
+  --parentscol P   column with the IDs of the parents (1-based, default: 2)
+  --commentchar C  lines starting with this character are ignored (default: #)
+
+Module input:
+  --module M  select file <M> as a Python module defining a function
+              yielding IDs of elements and parents
+  <args>...   arguments to be passed to the function;
+              if they contain a '=', they are passed as keyword arguments
+  --fn FN     name of the function to be called (default: element_parent_ids)
+
+Further options:
+  --force      overwrite the output file if it exists (action --new)
+  --debug      print debug information
+  --changes    verbosily report changes made to the tree
+  --quiet      disable log messages
+  --help       show this help message and exit
+  --version    show program's version number and exit
+"""
+
+from docopt import docopt
+from pathlib import Path
+from fastsubtrees import Tree, logger, _scripts_support, VERSION, attribute
+
+NCBI_DUMP_SEP = "\t|\t"
+NCBI_DUMP_TAXID_COL = 0
+NCBI_DUMP_PARENT_COL = 1
+
+def get_generator(args):
+  fn = "element_parent_ids"
+  if args["--module"]:
+    if args["--ncbi"]:
+      logger.warning("The --ncbi option is ignored when using --module")
+    if args["--fn"]:
+      fn = args["--fn"]
+    logger.debug("Using function {} from module {} as a source of IDs".\
+        format(fn, args["--module"]))
+    m = _scripts_support.get_module(args["--module"], fn)
+    posargs, keyargs = _scripts_support.get_fn_args(True, args["<args>"])
+  else:
+    if args["--fn"]:
+      logger.warning("Argument --fn ignored when not using --module")
+    import fastsubtrees.ids_modules.ids_from_tabular_file as m
+    logger.debug("Using tabular file {} as a source of IDs".\
+        format(args["<tabfile>"]))
+    posargs = [args["<tabfile>"]]
+    if args["--ncbi"]:
+      if args["--separator"] or args["--elementscol"] \
+          or args["--parentscol"] or args["--commentchar"]:
+        logger.warning("The --ncbi option overrides the following options: " +
+            "--separator, --elementscol, --parentscol, --commentchar")
+      keyargs = {"separator": NCBI_DUMP_SEP,
+                 "element_id_column": NCBI_DUMP_TAXID_COL,
+                 "parent_id_column": NCBI_DUMP_PARENT_COL,
+                 "comment_pfx": "#"}
+    else:
+      keyargs = {"separator": "\t",
+                 "element_id_column": 0,
+                 "parent_id_column": 1,
+                 "comment_pfx": "#"}
[…]
