+++ fastsubtrees/commands/query.py
@@ -29,7 +29,7 @@ Options:
 
 import os
 from docopt import docopt
-from fastsubtrees import Tree, logger, _scripts_support, VERSION, attribute
+from fastsubtrees import Tree, logger, _scripts_support, __version__
 
 def check_attribute_args(args):
   attrnames = args["<attribute>"]
@@ -50,49 +50,34 @@ def check_attribute_args(args):
     args["--show-none"] = True
   return attrnames
 
-def get_attribute_values(attrnames, tree, subtree_root, args):
-  attr_values = {}
-  for attrname in attrnames:
-    attributefile = attribute.attrfilename(args["<tree>"], attrname)
-    if not os.path.isfile(attributefile):
-      logger.error("Attribute '{}' not found".format(attrname))
-      exit(1)
-    logger.debug("Loading attribute '{}' values from file '{}'".\
-        format(attrname, attributefile))
-    attr_values[attrname] = attribute.get_attribute_list(tree, subtree_root, \
-        attributefile)
-    if args["--stats"]:
-      filtered = [a for a in attr_values[attrname] if a is not None]
-      flattened = [e for sl in filtered for e in sl]
-      logger.info("Number of nodes with attribute '{}': {}".\
-          format(attrname, len(filtered)))
-      logger.info("Number of values of attribute '{}': {}".\
-          format(attrname, len(flattened)))
-  return attr_values
+def show_header(args, attrnames):
+  if not args["--no-header"]:
+    header_data = []
+    if not args["--attributes_only"]:
+      header_data.append("node_id")
+    if args["--subtree-sizes"]:
+      header_data.append("subtree_size")
+    if args["--parents"]:
+      header_data.append("parent")
+    header_data.extend(attrnames)
+    print("# "+"\t".join(header_data))
 
-def show_header(attrnames, attributes_only):
-  header_data = []
-  if not attributes_only:
-    header_data.append("node_id")
-  header_data.extend(attrnames)
-  print("# "+"\t".join(header_data))
-
-def show_results(args, tree, node_ids, attr_values, attrnames):
+def show_data(args, subtree_info, attrnames):
   n_nodes = 0
   if not args["--separator"]:
     args["--separator"] = "\t"
-  for i, node_id in enumerate(node_ids):
-    if node_id == tree.DELETED:
+  for i, node_id in enumerate(subtree_info["node_id"]):
+    if node_id == Tree.DELETED:
       continue
     n_nodes += 1
     if not args["--show-none"] and \
-        all([attr_values[attrname][i] is None for attrname in attrnames]):
+        all([subtree_info[attrname][i] is None for attrname in attrnames]):
       continue
     line_data = []
     if not args["--attributes-only"]:
       line_data.append(str(node_id))
     for attrname in attrnames:
-      value = attr_values[attrname][i]
+      value = subtree_info[attrname][i]
       if isinstance(value, list):
         line_data.append(", ".join([str(v) for v in value]))
       else:
@@ -100,21 +85,6 @@ def show_results(args, tree, node_ids, attr_values, attrnames):
     print(args["--separator"].join(line_data))
   if args["--stats"]:
     logger.info("Number of nodes in subtree: {}".format(n_nodes))
-
-def get_parents(tree, node_ids, subtree_root):
-  parents = []
-  for node_id in node_ids:
-    if node_id == subtree_root:
-      parents.append(node_id)
-    else:
-      parents.append(tree.get_parent(node_id))
-  return parents
-
-def get_subtree_sizes(tree, node_ids):
-  subtree_sizes = []
-  for node_id in node_ids:
[…]
