+++ fastsubtrees/commands/datatype.md
@@ -1,4 +1,0 @@
-  --datatype DT  function to apply to the attribute values before writing
-                 them to the output file; <DT> must be a valid Python function;
-                 the function can optionally be defined in the module specified
-                 by <attrmod> [default: str]

[…]
