### 2022-10-15 de538e3
create cov report when running make testcov
