+++ tests/test_cli.py
@@ -56,15 +56,153 @@ def test_query(testout, testdata, script, script_runner,
           testdata("small_tree_attrX.tsv")]
   ret = script_runner.run(script("fastsubtrees"), *args)
   # query the constructed tree
-  args = ["query", small_tree_file, "8", "-N"]
+  args = ["query", small_tree_file, "8", "-H"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout == "".join(f"{x}\n" for x in results_query_small_tree_id_8)
   # query the attribute
-  args = ["query", small_tree_file, "8", "attrX", "-a", "-N"]
+  args = ["query", small_tree_file, "8", "attrX", "-a", "-H"]
   ret = script_runner.run(script("fastsubtrees"), *args)
   assert ret.returncode == 0
   assert ret.stdout.strip() == str(results_query_small_tree_id_8_attrX)
+
+def attribute_values_wrapper(filename, a, b=None):
+  assert (a == "A")
+  assert (b == "B=B")
+  import fastsubtrees.ids_modules.attr_from_tabular_file as aftf
+  return aftf.attribute_values(filename)
+
+def attribute_values_wrapperB(filename, a, b=None):
+  assert (a == "b=B=B")
+  assert (b == "A")
+  import fastsubtrees.ids_modules.attr_from_tabular_file as aftf
+  return aftf.attribute_values(filename)
+
+@pytest.mark.script_launch_mode('subprocess')
+def test_attribute_failures_and_warnings(testout, testdata, ids_modules, script,
+    script_runner, small_tree_file):
+  # tree file does not exist
+  args = ["attribute", "does_not_exist.tree", 'attrY', "--add",
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+  # module does not exist
+  args = ["attribute", small_tree_file, 'attrX', "--module", "not_existing.py",
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+  # modules does not define the function
+  args = ["attribute", small_tree_file, 'attrX', "--fn", "not_existing_fn",
+          "--module", ids_modules("ids_from_tabular_file.py"),
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+  # modules cannot be loaded
+  args = ["attribute", small_tree_file, 'attrX',
+          "--module", testdata("small_tree_attrX.tsv"),
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+  # using --fn without --module, produces a warning, but no error
+  args = ["attribute", small_tree_file, 'attrX', "--fn", "attribute_values",
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  assert "WARNING" in ret.stderr
+  # attribute file not existing
+  args = ["attribute", small_tree_file, 'attrY', "--add",
+          testdata("small_tree_attrX.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+  # delete non-existing node, no --strict, no error
[…]
