### 2022-10-17 f4ed587
update GAV app to new attribute creation API
