### 2022-10-17 dcd0009
update ntsubtree to the new API
