### 2022-10-17 6d599f4
update benchmarks and docker scripts
