+++ benchmarks/benchmarks_sql.sh
@@ -4,17 +4,21 @@
 # Benchmarks recursive SQL based subtree extraction
 #
 
-if [ $# -ne 4 ]; then
-    echo "Usage: $0 <dbuser> <dbpass> <dbname> <dbsock>"
-    echo "where <dbuser> is the database user name, <dbpass> is the"
-    echo "database password, <dbname> is the database name, and <dbsock>"
-    echo "is the database socket."
+if [ $# -ne 5 ]; then
+    echo "Usage: $0 <dbuser> <dbpass> <dbname> <dbsock> <ntdumpsdir>"
+    echo "where:"
+    echo "  <dbuser> is the database user name"
+    echo "  <dbpass> is the database password"
+    echo "  <dbname> is the database name"
+    echo "  <dbsock> is the database socket file"
+    echo "  <ntdumpsdir> is a directory containing NCBI taxonomy dump files"
     exit 1
 fi
 DBUSER=$1
 DBPASS=$2
 DBNAME=$3
 DBSOCK=$4
+NTDUMPSDIR=$5
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 FST_DIR=$SCRIPT_DIR/..
@@ -31,7 +35,7 @@ for ((i=0; i<$NREPEATS; i++)); do
   echo "Step $STEP, iteration $i..."
   /usr/bin/time -f "$STEP\t$ROOT\t$i\t%U\t%S\t%e\t%M" -o $OUTFILE -a \
     ntmirror-dbload --testmode --reset \
-      $DBUSER $DBPASS $DBNAME $DBSOCK ntdumpdir
+      $DBUSER $DBPASS $DBNAME $DBSOCK $NTDUMPSDIR
 done
 for ((i=0; i<$NREPEATS; i++)); do
   STEP="extract"
