### 2022-10-17 d6db735
recompute benchmarks with newest version
