### 2022-10-17 a82df4e
add a failing case and fix it
