### 2022-10-17 6028f59
make subtree_size contain the size INCLUDING the node
