### 2022-10-17 7f0ec39
unify Tree.DELETED and Tree.UNDEF
