### 2022-10-17 a11d9cd
use UNDEF as default value for treedata
