### 2022-10-18 762357f
greatly improve efficiency of tree construction
