### 2022-10-18 a5defbe
improve memory effiency of fast construction algorithm
