### 2022-10-18 aac5544
recompute benchmarks, run all benchmarks by default
