+++ docs/paper.md
@@ -104,11 +104,11 @@ dynamic, i.e. it is possible to add and remove subtrees.
 
 # Subtree extraction benchmarks
 
-Benchmarks were performed on a MacBook Pro 2021 with Apple M1 Pro CPU
-and 32 Gb RAM. Running times were measured as an average of 3 runs using
-GNU time version 1.9 [@GNUtime].
-For the tests Python version 3.10.2 was used.
-The tested version of fastsubtrees was 2.0.
+Benchmarks were performed on a MacBook Pro 2021 with Apple M1 Pro CPU and 32 Gb
+RAM. Running times were measured as an average of 3 runs using GNU time version
+1.9 [@GNUtime]. Thereby, CPU time was computed as the sum of user and system
+time. For the tests Python version 3.10.2 was used. The tested version of
+fastsubtrees was 2.0.
 
 The NCBI taxonomy tree used for the tests was downloaded on October 7, 2022
 from the NCBI FTP website [@NCBI:FTP].
@@ -119,16 +119,17 @@ of the source code repository, version 1.4)
 The tree contained 2447574 nodes. The generation of the tree
 representation of the NCBI taxonomy tree from the dump files
 using the _fastsubtrees construct_ command required
-9 minutes and 23 seconds (average of 3 runs).
+12.5 seconds (average of 3 runs).
 
 An alternative to the use of _fastsubtrees_ is to store the tree data in a SQL
 database and extract subtrees using hierarchical SQL queries. We implemented
-this solution in a package _ntmirror_: the dump data downloaded by _ntdownload_
-is loaded into a MariaDB database (version 10.6.10)
-and the script _ntmirror-extract-subtree_, based on
-SQLAlchemy allows to extract a subtree using SQL.
-The installation of _ntmirror_ using ``pip`` requires the installation
-of MariaDB and its Python connector (_mariadb_ package).
+this solution in a package _ntmirror_. First, the dump data downloaded by
+_ntdownload_ is loaded into a MariaDB database (version 10.6.10) using the
+script _ntmirror-dbload_, requiring 28.8 seconds (average of 3 runs).
+Thereafter, subtrees are extracted using the script _ntmirror-extract-subtree_,
+based on SQLAlchemy, implementing a hierarchical SQL query. The installation of
+_ntmirror_ using ``pip`` requires the installation of MariaDB and its Python
+connector (_mariadb_ package).
 
 To select subtrees of different sizes for the benchmarks, we started from the
 taxonomy ID of _Escherichia coli_ K12 MG1655 (511145) and climbed up the
@@ -137,19 +138,22 @@ node (TaxID 2), i.e. nodes 83333 (_Escherichia coli_ K12),
 562 (_Escherichia coli_), 561 (_Escherichia_ genus), 543 (Enterobacteriaceae),
 91347 (Enterobacterales), 1236 (Gammaproteobacteria) and 1224 (Proteobacteria).
 The running time and memory usage of _fastsubtree_ are compared with those for
[…]
