### 2022-10-18 79a687b
make attribute editing API clearer
