### 2022-10-18 4afb821
better organize commands code
