### 2022-10-18 9ad6fb3
ntsubtree: implement CLI and update to fastsubtrees 2
