+++ ntsubtree/setup.py
@@ -10,7 +10,7 @@ def readme():
     return f.read()
 
 setup(name='ntsubtree',
-      version='0.1.1',
+      version='1.0',
       #description='Tree representation for fast queries of '+\
       #            'the subtree of a taxon in the NCBI taxonomy tree',
       description='preliminary version',
