+++ ntsubtree/ntsubtree/constants.py
@@ -10,9 +10,6 @@ __author__="ggonnella"
 NCBI_DUMP_SEP = "\t|\t"
 
 NCBI_NODES_DUMP_FILENAME = "nodes.dmp"
-NCBI_NODES_DUMP_TAXID_COL = 0
-NCBI_NODES_DUMP_PARENT_COL = 1
-
 NCBI_NAMES_DUMP_FILENAME = "names.dmp"
 NCBI_NAMES_DUMP_TAXID_COL = 0
 NCBI_NAMES_DUMP_NAME_COL = 1
