### 2022-10-18 e8c97b9
attribute --list option in CLI
