### 2022-10-18 24a26ed
improve/fix ntsubtree commands
