### 2022-10-18 0ecd381
make code more modular
