+++ fastsubtrees/tree.py
@@ -13,8 +13,11 @@ from pathlib import Path
 from fastsubtrees import logger, tqdm, error
 from fastsubtrees.ids_modules import ids_from_tabular_file, \
                                      attr_from_tabular_file
+from .attribute import TreeAttributes
+from .query import SubtreeQuery
+from .edit import TreeEditor
 
-class Tree():
+class Tree(TreeAttributes, SubtreeQuery, TreeEditor):
 
   def __init__(self):
     # treedata contains:
@@ -236,7 +239,7 @@ class Tree():
     self.filename = filename
     return self
 
-  def __check_node_number(self, node):
+  def _check_node_number(self, node):
     if node < 0 or node > len(self.coords) - 1:
       raise error.NodeNotFoundError(f"Node ID '{node}' does not exist.")
 
@@ -244,369 +247,8 @@ class Tree():
     """
     Returns the parent ID of the given node.
     """
-    self.__check_node_number(node)
-    if node == self.root_id:
-      return node
-    else:
-      return self.parents[node]
-
-  def get_subtree_size(self, node: int) -> int:
-    """
-    Returns the number of nodes in the subtree rooted at the given node.
-    """
-    self.__check_node_number(node)
-    return self.subtree_sizes[node]
-
-  def get_treedata_coord(self, node: int) -> int:
-    """
-    Returns the position of the given node in the treedata array.
-    """
-    self.__check_node_number(node)
-    return self.coords[node]
-
-  def get_subtree_data(self, subtree_root: int) -> array.array:
-    """
-    Returns the treedata array for the subtree rooted at the given node.
-
-    This includes nodes marked as deleted.
-    """
-    self.__check_node_number(subtree_root)
-    pos = self.get_treedata_coord(subtree_root)
-    if pos > 0:
-      subtree_size = self.get_subtree_size(subtree_root)
-      return self.treedata[pos:pos + subtree_size]
-    else:
-      return array.array("Q")
-
-  def subtree_ids(self, subtree_root: int) -> array.array:
-    """
-    Returns the IDs of the nodes in the subtree rooted at the given node.
-    """
-    self.__check_node_number(subtree_root)
-    result = array.array("Q")
-    for node_id in self.get_subtree_data(subtree_root):
-      if node_id != self.UNDEF:
-        result.append(node_id)
-    return result
-
-  NONELINE = 'null\n'
-
-  def __edit_attribute_values(self, edit_script, attrfilenames):
-    for attrfilename in attrfilenames:
-      with open(attrfilename, 'r') as f:
-        lines = f.readlines()
-      for op in edit_script:
-        if op[0] == "insert":
-          lines.insert(op[1]-1, Tree.NONELINE)
-        elif op[0] == "copy":
-          lines[op[1]-1] = lines[op[2]-1]
-        elif op[0] == "delete":
-          lines[op[1]-1] = Tree.NONELINE
-      with open(attrfilename, 'w') as f:
-        f.writelines(lines)
-
-  def __add_or_move_nodes(self, generator, attrfilenames=[],
-      skip_existing=False, rm_existing_set=None, list_added=None,
-      list_moved=None, total=None, edit_script=None):
-    n_added = 0
-    n_moved = 0
-    pending = defaultdict(list)
-    logger.info("Tree root: " + str(self.root_id))
-    if edit_script is None:
-      edit_script = []
-    for node_number, parent in tqdm(generator, total=total):
-      if rm_existing_set is not None:
-        rm_existing_set.discard(node_number)
-      if node_number < 0:
-        raise error.ConstructionError(\
-            f"The node IDs must be >= 0, found: {node_number}")
-      elif parent < 0:
-        raise error.ConstructionError(\
[…]
