+++ fastsubtrees/edit.py
@@ -1,0 +1,319 @@
+"""
+Editing operations for fastsubtrees.Tree objects.
+"""
+
+import struct
+import array
+import sys
+from collections import defaultdict
+from typing import List, Union, Iterator, Tuple, Dict, Any
+import json
+import glob
+from pathlib import Path
+from fastsubtrees import logger, tqdm, error
+from fastsubtrees.ids_modules import ids_from_tabular_file, \
+                                     attr_from_tabular_file
+
+class TreeEditor():
+
+  def __add_or_move_nodes(self, generator, attrfilenames=[],
+      skip_existing=False, rm_existing_set=None, list_added=None,
+      list_moved=None, total=None, edit_script=None):
+    n_added = 0
+    n_moved = 0
+    pending = defaultdict(list)
+    logger.info("Tree root: " + str(self.root_id))
+    if edit_script is None:
+      edit_script = []
+    for node_number, parent in tqdm(generator, total=total):
+      if rm_existing_set is not None:
+        rm_existing_set.discard(node_number)
+      if node_number < 0:
+        raise error.ConstructionError(\
+            f"The node IDs must be >= 0, found: {node_number}")
+      elif parent < 0:
+        raise error.ConstructionError(\
+            f"The node parent IDs must be >= 0, found: {parent}")
+      if node_number < len(self.parents):
+        if node_number == self.root_id:
+          if skip_existing:
+            if node_number == parent:
+              continue
+            else:
+              raise error.ConstructionError(\
+                  f'Node {node_number} / parent {parent} already exists '+\
+                  f'as the root node')
+          else:
+            raise error.ConstructionError(\
+                f"The root node {node_number} already exists")
+        if self.parents[node_number] != self.UNDEF and \
+              self.treedata[self.coords[node_number]] != self.UNDEF:
+          if skip_existing:
+            if self.parents[node_number] != parent:
+              if parent >= len(self.parents) or self.coords[parent] == 0:
+                pending[parent].append(('move', node_number))
+              else:
+                self.__move_subtree(node_number, parent, edit_script)
+                n_moved += 1
+                if list_moved is not None:
+                  list_moved.append(node_number)
+            continue
+          else:
+            if self.parents[node_number] != parent:
+              raise error.ConstructionError(\
+                f"Node {node_number} / parent {parent} already exists "+\
+                f"with a different parent ({self.parents[node_number]})")
+            else:
+              raise error.ConstructionError(\
+                f"Node {node_number} / parent {parent} already exists")
+      if parent >= len(self.parents) or self.coords[parent] == 0:
+        pending[parent].append(("insert", node_number))
+      else:
+        self.__insert_node(node_number, parent, edit_script, list_added)
+        n_added += 1
+        stack = [node_number]
+        while stack:
+          node = stack.pop()
+          if node in pending:
+            for child in pending[node]:
+              if child[0] == "insert":
+                self.__insert_node(child[1], node, edit_script, list_added)
+                n_added += 1
+                stack.append(child[1])
+              elif child[0] == "move":
+                self.__move_subtree(child[1], node, edit_script)
+                n_moved += 1
+                if list_moved is not None:
+                  list_moved.append(child[1])
+            del pending[node]
+    if len(pending) > 0:
+      raise error.ConstructionError(\
+          "Impossible operations because the node parents " + \
+          f"were not present in the tree: {pending}")
+    self._edit_attribute_values(edit_script, attrfilenames)
[…]
