+++ fastsubtrees/commands/query.py
@@ -30,7 +30,7 @@ Options:
 
 import os
 from docopt import docopt
-from fastsubtrees import Tree, logger, _scripts_support, __version__
+from fastsubtrees import Tree, logger, __version__
 
 def check_attribute_args(args):
   attrnames = args["<attribute>"]
