### 2022-10-18 55a14f1
reset operation, sometimes faster than update
