+++ tests/test_edit_tree.py
@@ -93,28 +93,116 @@ def test_delete_node_err_not_exist(testdata):
   with pytest.raises(error.NodeNotFoundError):
     tree.delete_subtree(99)
 
-def test_update_tree(testdata):
+def assert_identical_subtrees(tree1, tree2):
+  for node in set(tree1.subtree_ids(tree1.root_id)):
+    assert set(tree1.subtree_ids(node)) == set(tree2.subtree_ids(node))
+
+def assert_identical_attributes(tree1, tree2):
+  attrnames1 = set(tree1.list_attributes())
+  attrnames2 = set(tree2.list_attributes())
+  assert attrnames1 == attrnames2
+  for attrname in attrnames1:
+    attrvalues1 = tree1.load_attribute_values(attrname)
+    attrvalues2 = tree2.load_attribute_values(attrname)
+    assert attrvalues1 == attrvalues2
+
+def test_update_tree(testdata, testout):
   tree = Tree.construct_from_tabular(testdata('small_tree.tsv'))
+  tree.to_file(testout('small_tree.tree'))
+  tree.destroy_all_attributes()
+  tree.create_attribute_from_tabular("attrX", testdata("small_tree_attrX.tsv"))
+  tree.create_attribute_from_tabular("attrX2", testdata("small_tree_attrX.tsv"))
   tree.update_from_tabular(testdata('small_tree.update.tsv'))
   with open(testdata("small_tree.updated.query.root.parents.results")) as f:
     expected_results = [int(line.split("\t")[0]) for line in f if line[0] != "#"]
   assert list(tree.subtree_ids(tree.root_id)) == expected_results
+  t2 = Tree.construct_from_tabular(testdata('small_tree.update.tsv'))
+  t2.to_file(testout('small_tree2.tree'))
+  t2.destroy_all_attributes()
+  t2.create_attribute_from_tabular("attrX", testdata("small_tree_attrX.tsv"))
+  t2.create_attribute_from_tabular("attrX2", testdata("small_tree_attrX.tsv"))
+  assert_identical_subtrees(tree, t2)
+  assert_identical_attributes(tree, t2)
+  # update deleting everything except root
   tree.update_from_tabular(testdata('small_tree.rootonly.tsv'))
   assert list(tree.subtree_ids(tree.root_id)) == [1]
+  t2 = Tree.construct_from_tabular(testdata('small_tree.rootonly.tsv'))
+  t2.to_file(testout('small_tree2.tree'))
+  t2.destroy_all_attributes()
+  t2.create_attribute_from_tabular("attrX", testdata("small_tree_attrX.tsv"))
+  t2.create_attribute_from_tabular("attrX2", testdata("small_tree_attrX.tsv"))
+  assert_identical_subtrees(tree, t2)
+  assert_identical_attributes(tree, t2)
+  # update_from_ncbi_dump
   tree = Tree.construct_from_ncbi_dump(testdata('small_ncbi.tsv'))
+  tree.to_file(testout('small_ncbi.tree'))
+  tree.destroy_all_attributes()
+  tree.create_attribute_from_tabular("attrX", testdata("small_tree_attrX.tsv"))
+  tree.create_attribute_from_tabular("attrX2", testdata("small_tree_attrX.tsv"))
   tree.update_from_ncbi_dump(testdata("small_ncbi.update.tsv"))
   with open(testdata("small_ncbi.updated.query.root.parents.results")) as f:
     expected_results = [int(line.split("\t")[0]) for line in f if line[0] != "#"]
   assert list(tree.subtree_ids(tree.root_id)) == expected_results
[…]
