+++ docs/API.md
@@ -84,16 +84,28 @@ will be automatically adapted.
 The method will raise an exception if the new parent node is not present
 in the tree.
 
-### Updating a tree
+### Updating or resetting a tree
 
 A tree can be updated to reflect the contents of a data source of
-tuples ``(node_id, parent)``. Thus all nodes which are present in the
+tuples ``(node_id, parent)``. Thereby all nodes which are present in the
 tree but not generated by the data source, will be removed, nodes
 with a changed parent will be moved, and nodes not present in the
 tree will be added.
-
+The method for updating a tree is ``tree.update(generator)``.
 The method will fail if the root is changed, or if the tree data
 is invalid.
+
+Updating can be advantageous over creating a new tree, in terms of
+performance, if the changes are not many.
+If a tree had several changes, a more efficient way is to recreate
+the tree data, using the ``tree.reset(generator)`` method.
+This is equivalent to creating a new tree, but the attribute values
+of nodes are kept, if the nodes are still present in the tree
+after the reset.
+
+Both update and reset have a ``_from_tabular`` and a ``_from_ncbi_dump``
+version of the method for working with tabular files and, respectively,
+the NCBI taxonomy nodes dump file.
 
 ## Working with attributes
 

[…]
