### 2022-10-18 d5d67da
reset operation for the CLI
