+++ tests/test_cli.py
@@ -80,6 +80,42 @@ def test_update(testout, testdata, script, script_runner):
   assert "ERROR" in ret.stderr
 
 @pytest.mark.script_launch_mode('subprocess')
+def test_reset(testout, testdata, script, script_runner):
+  Path(testout("small_tree.tree")).unlink(missing_ok=True)
+  args = ["tree", testout("small_tree.tree"), testdata("small_tree.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  # updating the tree with itself
+  args = ["tree", testout("small_tree.tree"), "--reset",
+      testdata("small_tree.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  # really updating now
+  args = ["tree", testout("small_tree.tree"), "--reset",
+      testdata("small_tree.update.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  with open(testdata("small_tree.updated.query.root.parents.results")) as f:
+    expected_results = f.read()
+  args = ["query", testout("small_tree.tree"), "root", "--parents"]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  assert set(ret.stdout.split("\n")) == set(expected_results.split("\n"))
+  args = ["query", testout("small_tree.tree"), "root", "--subtree-sizes"]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  # updating to just the root
+  args = ["tree", testout("small_tree.tree"), "--reset",
+      testdata("just_root.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 0
+  # try to set root node as internal
+  args = ["tree", testout("small_tree.tree"), "--reset",
+      testdata("same_node_root_and_not.tsv")]
+  ret = script_runner.run(script("fastsubtrees"), *args)
+  assert ret.returncode == 1
+  assert "ERROR" in ret.stderr
+
+@pytest.mark.script_launch_mode('subprocess')
 def test_new_from_nodes_dmp(testout, testdata, script, script_runner):
   Path(testout("small_ncbi.tree")).unlink(missing_ok=True)
   args = ["tree", testout("small_ncbi.tree"), testdata("small_ncbi.tsv"),
