+++ ntsubtree/ntsubtree/__init__.py
@@ -1,5 +1,5 @@
 # NtSubtree package
 
 from .constants import __version__
-from .ntsubtree import __auto_setup, setup, update, get_tree
+from .ntsubtree import __auto_setup, setup, update, get_tree, search_name
 __auto_setup()
