### 2022-10-18 272a401
mention NtSubtree in README and update documentation of sub-packages
