+++ README.md
@@ -1,67 +1,43 @@
 # Fastsubtrees
 
-Fastsubtrees is a Python library and a command line script, for handling fairly
+_Fastsubtrees_ is a Python library and a command line script, for handling fairly
 large trees (in the order of magnitude of millions nodes), in particular
 allowing the fast extraction of any subtree.
-
-The main application domain of fastsubtrees is working with the NCBI taxonomy
+The main application domain of _fastsubtrees_ is working with the NCBI taxonomy
 tree, however the code is implemented in a generic way, so that other
 applications are possible.
 
-For achieving an efficient running time and memory use, the nodes of the tree
-are represented compactly in deep-first traversal order.
-Subtrees are then extracted in O(s) time, where s is the size of the extracted
-subtree (i.e. not depending on the size of the whole tree).
+The library functionality can be accessed both from inside Python code
+and from the provided command line tool ``fastsubtrees``.
 
-The tree representation can be saved to file, so that it must be not be
-re-computed each time. It is dynamical, i.e. after a tree has been created,
-it can be modified, by adding a new leaf node,
-or an entire subtree under an existing node. Also, existing leaf nodes or
-entire subtrees can be deleted.
+## Introduction
 
-Along with numerical node identifiers, additional
-information can be stored for each node, by defining any number of optional
-node attributes.
+For the use of _fastsubtrees_, nodes must be uniquely identified by non-negative IDs.
+Furthermore, the space of the IDs must be compact (i.e. the maximum ID should not be
+much larger than the number of IDs).
 
-## Node identifiers
+The first step when using _fastsubtrees_ is to construct a tree representation.
+The operation requires a source of IDs of elements and their parents, which can be
+a tabular file, or any Python function yielding the IDs.
 
-Each node must be represented by a unique non-negative ID. The IDs must not
-necessarily be all consecutive (i.e. some "holes" may be present), but the
-largest node ID (_idmax_) should not be much larger than the total number of
-nodes, because the memory consumption is in _O(idmax)_.
+This operation just takes a few seconds, for a tree with million nodes, such as the NCBI taxonomy tree.
+It must be done only once, if a tree does not change, since the resulting data
+is stored to file.
 
-The NCBI taxonomy tree, for example, fullfills the conditions stated above.
+The IDs of the NCBI taxonomy tree fullfill the conditions stated above. However, the library
+can be used for any tree. A way to use the library with IDs which do not fullfill the conditions,
+it to map them to an ID space which does, and store the original IDs as an attribute.
 
-However, the library can be used on any other tree. For this, if the IDs are
-non-numerical, contain negative numbers, or the ID space is not compact,
-they must be first mapped to different IDs.
+Besides the IDs, a tree can contain further information, e.g. integers, floats or other
+data, here called attributes, associated to the nodes. Each node can contain zero, one or more values
+for an attribute. To add values for an attribute, a tabular file or another data
+source (a Python function) is selected.
 
-## Attributes
+The data for any subtree can then be easily and efficently queried; thereby the node IDs and/other
+selected attributes can be retrieved.
 
-Besides a numerical node identifier, for each node of the tree additional
-information can be stored in form of _attributes_.  Any number of attributes
-can be created.  Each node can contain a single value or a set of values for an
[…]
