### 2022-10-19 186b831
ntdownload: expose ntnames to API and complete test suite
