+++ ntdownload/tests/conftest.py
@@ -10,6 +10,10 @@ def testout():
   return lambda fn: Path(__file__).parent / 'output' / fn
 
 @pytest.fixture
+def testdata_dir():
+  return Path(__file__).parent / 'testdata'
+
+@pytest.fixture
 def script():
   return lambda fn: Path(__file__).parent.parent / 'bin' / fn
 
