+++ ntdownload/tests/testdata/README.md
@@ -1,0 +1,2 @@
+These files are the first 10 lines of the original files in the NCBI taxonomy database.
+They are used for test purposes only.
