### 2022-10-19 815bee3
make docker-benchmarks automatically output paper tables
