### 2022-10-19 7e9a8ba
install also the other sub-packages in Docker
