### 2022-10-19 ggonnella/fastsubtrees#16 @ggonnella
The new version contains an almost complete coverage of the entire fastsubtrees code, which can be shown using ``make testcov``.

The current output is:
```
---------- coverage: platform darwin, python 3.9.12-final-0 ----------
Name                                                 Stmts   Miss  Cover
bin/fastsubtrees                                        22      0   100%
fastsubtrees/__init__.py                                15      0   100%
fastsubtrees/attribute.py                              184      0   100%
fastsubtrees/commands/__init__.py                        0      0   100%
fastsubtrees/commands/_support.py                       56      0   100%
fastsubtrees/commands/attribute.py                      62      0   100%
fastsubtrees/commands/query.py                          70      0   100%
fastsubtrees/commands/tree.py                           82      0   100%
fastsubtrees/edit.py                                   205      0   100%
fastsubtrees/error.py                                   16      0   100%
fastsubtrees/ids_modules/__init__.py                     0      0   100%
fastsubtrees/ids_modules/attr_from_tabular_file.py      16      0   100%
fastsubtrees/ids_modules/ids_from_database.py           12     12     0%
fastsubtrees/ids_modules/ids_from_tabular_file.py       18      0   100%
fastsubtrees/query.py                                   52      0   100%
fastsubtrees/tree.py                                   152      0   100%
TOTAL                                                  962     12    99%
```

The ids_from_database is not tested; it is an example illustrating how to use a database table as a data source, it is not used in anything in the paper, benchmarks or example app, and including in the tests would again return to the dependency to a database, which I want to avoid.
