### 2022-10-19 ggonnella/fastsubtrees#8 @ggonnella
I greatly improved the performance by understanding where the bottleneck of the tree construction was located.

In particular, it needed most of the time in computing the next free position in a subtree, while constructing it. Now I used the coords array to store the next free position during construction (and computed the final coordinates at the end by subtracting from each value the subtree size).

The new version just takes about 15 seconds instead of 10 minutes to constructing the tree.
Thus a parallel construction (which would be quite complicated to realize) is much less interesting.
