+++ fastsubtrees/query.py
@@ -2,17 +2,8 @@
 Subtree query operations for Tree() objects.
 """
 
-import struct
 import array
-import sys
-from collections import defaultdict
-from typing import List, Union, Iterator, Tuple, Dict, Any
-import json
-import glob
-from pathlib import Path
-from fastsubtrees import logger, tqdm, error
-from fastsubtrees.ids_modules import ids_from_tabular_file, \
-                                     attr_from_tabular_file
+from typing import List, Dict, Any
 
 class SubtreeQuery():
 
