### 2022-10-19 ggonnella/fastsubtrees#16 @ggonnella
Update: I decided to remove the ``fastsubtrees/ids_modules/ids_from_database.py`` completely.
