+++ ntmirror/tests/test_dbload_cli.py
@@ -33,6 +33,7 @@ def test_dbload_cli_sqlalchemy(connection, script, testdatadir,
     args = connection_args + [testdatadir, "--exitcode", "--sqlalchemy",
                                            "--dbecho", "--verbose", "--testmode"]
     ret = script_runner.run(script("ntmirror-dbload"), *args)
+    assert ret.returncode == 0
     session = Session(connection)
     for klass in dbschema.DB_MODEL_CLASSES:
       assert session.query(klass).count() == 10
