### 2022-10-19 a699f94
avoid ntmirror tests for Python package action
