### 2022-10-19 ggonnella/fastsubtrees#17 @ggonnella
Linting with flakes8 and tests are now run using GitHub actions
