### 2022-10-19 56b8156
add test requirements to python-package.yml
