### 2022-10-19 76690e8
fix duplicate description in ntsubtree setup.py
