+++ ntsubtree/setup.py
@@ -12,7 +12,6 @@ setup(name='ntsubtree',
       version='1.0',
       description='Tree representation for fast queries of '+\
                    'the subtree of a taxon in the NCBI taxonomy tree',
-      description='preliminary version',
       long_description=readme(),
       long_description_content_type="text/markdown",
       url="htts://github.com/ggonnella/fastsubtrees/tree/main/ntsubtree",
