### 2022-10-19 debfc63
add ntdownload to requirements of fastsubtrees
this allows the testing of both using make tests of fastsubtrees
