+++ requirements.txt
@@ -3,3 +3,4 @@ loguru>=0.5.1
 docopt>=0.6.2
 schema>=0.7.4
 sh>=1.14.2
+ntdownload>=1.6
