### 2022-10-19 6e1cc30
install local package using pip before testing in Python package action
