### 2022-10-19 3aa6dbf
add index.rst for generating documentation
