+++ index.rst
@@ -1,0 +1,9 @@
+FastSubtrees
+------------
+
+.. toctree::
+   :glob:
+
+   README
+   docs/CLI
+   docs/API
