### 2022-10-19 e8476d9
fix relative paths in docs/index.rst
