+++ docs/index.rst
@@ -4,6 +4,6 @@ FastSubtrees
 .. toctree::
    :glob:
 
-   README
-   docs/CLI
-   docs/API
+   ../README
+   CLI
+   API
