### 2022-10-19 cbcf743
configure according to https://markdown-publishing-guide.readthedocs.io/en/latest/markdown-publishing-guide.html#introduction
