### 2022-10-19 622350a
cite Knuth TAOCP for tree definition
