+++ docs/paper.bib
@@ -52,15 +52,6 @@
   url = {https://github.com/ggonnella/fastsubtrees}
 }
 
-@misc{Black:1999,
-  author = {Paul E. Black},
-  title = {Definition: tree},
-  year = {2017},
-  publisher = {NIST},
-  journal = {Algorithms and Theory of Computation Handbook},
-  url = {https://xlinux.nist.gov/dads/HTML/tree.html}
-}
-
 @article{Mendler:2019,
     author = {Mendler, Kerrin and Chen, Han and Parks, Donovan H and Lobb, Briallen and Hug, Laura A and Doxey, Andrew C},
     title = "{AnnoTree: visualization and exploration of a functionally annotated microbial tree of life}",
@@ -130,4 +121,15 @@ doi = {10.1186/s13104-019-4577-5}
   volume   =  2020,
   month    =  jan,
   year     =  2020
+}
+
+@inbook{Knuth:1997,
+  title={The art of computer programming, Volume {I}: {F}undamental Algorithms, 3rd Edition},
+  author = {Donald Ervin Knuth},
+  isbn = {0-201-89683-4},
+  year = {1997},
+  publisher = {Addison-Wesley},
+  address = {Reading, MA, USA},
+  url = {https://www-cs-faculty.stanford.edu/~knuth/taocp.html},
+  pages = {308}
 }
