### 2022-10-19 #4755 @ggonnella
@vinisalazar @KonradHoeffner
The current version 2.0 addresses the concerns and comments raised until now.

In particular:
- it is much easier to install, as it does not require any external installation step
- a Docker image is provided for providing benchmarks against a database-based solution (and related tests)
- continuous integrations tasks (testing, linting) were added and automatized using GithubActions
- the documentation is now available and automatically updated on Read-The-Docs
- the example app was made easier to setup and install, made into an separate package, which is installable using pip and running with a single command
- as desired, a package (ntsubtree), installable by pip, has been made (and described in the README) which automatically installs the NCBI taxonomy tree, allows to easily update it, and makes it trivial to work with taxonomic names alongside the taxa IDs
- the test suite has been expanded and the test coverage computed using pytest-cov
- since the tree construction algorithm has been optimized, it is now much faster
- the definition of tree in the paper has been made refer to a print source (Knuth TAOCP book)
- typos in the paper have been fixed
