+++ fastsubtrees/commands/tree.py
@@ -39,7 +39,7 @@ Further options:
   -d, --debug      print debug information
   -h, --help       show this help message and exit
   -V, --version    show program's version number and exit
-"""
+""" # noqa
 
 from pathlib import Path
 from fastsubtrees import Tree, logger
