+++ genomes_attributes_viewer/setup.py
@@ -20,7 +20,8 @@ setup(name='genomes_attributes_viewer',
         "dash-bootstrap-components==1.0.2", "dash-core-components==2.0.0",
         "dash-html-components==2.0.0", "dash-table==5.0.0", "Werkzeug==2.0.0",
         "Flask==2.1.2", "pandas"],
-      url='https://github.com/ggonnella/fastsubtrees/tree/main/genomes_attributes_viewer',
+      url='https://github.com/ggonnella/fastsubtrees/"+\
+            "tree/main/genomes_attributes_viewer',
       keywords="ncbi taxonomy webapp dash fastsubtrees",
       author='Giorgio Gonnella',
       author_email='gonnella@zbh.uni-hamburg.de',

[…]
