+++ ntdownload/tests/test_downloader_api.py
@@ -16,17 +16,23 @@ from ntdownload import Downloader
 #
 USE_REAL_FILE = False
 
+if USE_REAL_FILE:
+  CONTAINED_FILENAME = "README.md"
+  ARCHIVE_FILENAME = Downloader.DUMPFILENAME
+else:
+  CONTAINED_FILENAME = Downloader.TESTFILENAME
+  ARCHIVE_FILENAME = CONTAINED_FILENAME + Downloader.ARCHIVE_SUFFIX
+
 def test_first_download(testout):
   outdir=testout("first_download")
   sh.rm(outdir, "-rf")
-  downloader = Downloader(str(outdir))
+  downloader = Downloader(outdir)
   if not USE_REAL_FILE:
     downloader.set_testmode()
   was_downloaded = downloader.run()
   assert was_downloaded
-  if not USE_REAL_FILE:
-    assert os.path.exists(outdir/downloader.DUMPFILENAME)
   assert os.path.exists(outdir/downloader.TIMESTAMP)
+  assert os.path.exists(outdir/CONTAINED_FILENAME)
 
 def test_no_newer_version(testout):
   outdir=testout("no_newer_version")
@@ -35,16 +41,16 @@ def test_no_newer_version(testout):
   sh.touch(outdir/Downloader.TIMESTAMP)
   future = time.time() + 3600
   os.utime(outdir/Downloader.TIMESTAMP, (future, future))
-  downloader = Downloader(str(outdir))
+  downloader = Downloader(outdir)
   downloader.set_testmode()
   was_downloaded = downloader.run()
   assert not was_downloaded
-  assert not os.path.exists(outdir/downloader.DUMPFILENAME)
+  assert not os.path.exists(outdir/CONTAINED_FILENAME)
 
 def test_newer_version(testout):
   outdir=testout("newer_version")
   sh.rm(outdir, "-rf")
-  downloader = Downloader(str(outdir))
+  downloader = Downloader(outdir)
   downloader.set_testmode()
   downloader.run()
   stinfo = os.stat(outdir/downloader.DUMPFILENAME)
@@ -52,9 +58,42 @@ def test_newer_version(testout):
   before_last_version = mtime - 18000
   os.utime(outdir/Downloader.TIMESTAMP, \
            (before_last_version, before_last_version))
-  os.unlink(outdir/downloader.DUMPFILENAME)
-  assert not os.path.exists(outdir/downloader.DUMPFILENAME)
+  os.unlink(outdir/CONTAINED_FILENAME)
+  assert not os.path.exists(outdir/CONTAINED_FILENAME)
   was_downloaded = downloader.run()
   assert was_downloaded
-  assert os.path.exists(outdir/downloader.DUMPFILENAME)
+  assert os.path.exists(outdir/CONTAINED_FILENAME)
 
+def test_no_unpack(testout):
+  outdir=testout("no_unpack")
+  sh.rm(outdir, "-rf")
+  downloader = Downloader(outdir)
+  if not USE_REAL_FILE:
+    downloader.set_testmode()
+  was_downloaded = downloader.run(unpack=False)
+  assert was_downloaded
+  assert os.path.exists(outdir/Downloader.TIMESTAMP)
+  assert os.path.exists(outdir/ARCHIVE_FILENAME)
+
+def test_force_https(testout):
+  outdir=testout("force_https")
+  sh.rm(outdir, "-rf")
+  downloader = Downloader(outdir)
+  if not USE_REAL_FILE:
+    downloader.set_testmode()
+  was_downloaded = downloader.run(force_https=True)
+  assert was_downloaded
+  assert os.path.exists(outdir/Downloader.TIMESTAMP)
[…]
