+++ docs/gav.md
@@ -1,0 +1,1 @@
+../genomes_attributes_viewer/README.md
