+++ docs/index.rst
@@ -7,4 +7,7 @@ FastSubtrees
    README.md
    CLI.md
    API.md
+   ntsubtree.md
+   gav.md
    ntdownload.md
+   ntmirror.md
