### 2022-10-20 44d4d60
improve documentation of NtMirror
