### 2022-10-20 7e83e2b
ntmirror: complete documentation
