### 2022-10-20 9d22dc3
make __init__.py file in ntdownload/tests only when needed for testcov
