+++ docs/index.rst
@@ -1,8 +1,6 @@
 FastSubtrees
 ------------
 
-:Authors: Giorgio Gonnella
-
 .. toctree::
    :glob:
 
