### 2022-10-20 5e349cd
small fixes in CLI manual
