### 2022-10-20 ggonnella/fastsubtrees#8 @KonradHoeffner
Thanks! Can you reference that paper and write a sentence on why that is not implemented in fastsubtrees? Do I understand correctly that this constructing tree data step is only needed to be done once and thus is not the core functionality of fastsubtrees, which is about extraction?
