### 2022-10-20 #4755 @vinisalazar
Hi,

Version 2.0 addresses all of my major concerns, but I do have some (minor) outstanding points:

* Something which I did not notice on my initial review is that there are no docstrings for classes. At the very least the `Tree` class should have an informative docstring explaining how to use the class, but ideally all public classes should have a brief description of what they do/how they work. [PEP 0257](https://peps.python.org/pep-0257/) offers a simple overview for docstring conventions.
* On ggonnella/fastsubtrees#15 I posted the traceback that I get when running tests from the repository root.
* When running the code on the README, I get this error:
```
Cell In [1], line 7
      5 from fastsubtrees import Tree
      6 infile = "ntdumpsdir/nodes.dmp"
----> 7 tree = Tree.construct_from_ncbi_dumps(infile)     # create the tree
      8 results = tree.subtree_ids(562)                   # retrieve subtree IDs
     10 attrtab="data/accession_taxid_attribute.tsv.gz"         # data file

AttributeError: type object 'Tree' has no attribute 'construct_from_ncbi_dumps'
```

Presumably it should be `Tree.construct_from_ncbi_dump`, without the s at the end? It is important to double check that any code snippets on the README run without any problems.
