### 2022-10-20 cb4e3c4
make tree construction optionally parallel
