### 2022-10-20 ggonnella/fastsubtrees#8 @ggonnella
Absolutely.

The tree construction algorithm used here is the following, where the input data consists of 2-tuples (element_id, parent_id) and the maximum node id m is not much larger than the number of ids n:
- the input data is iterated and a tables _P_ of parents by id is constructed: _P[element_id]=parent_id_, which requires _O(n)_ steps for reading the IDs and _O(m)_ steps for writing either the ID or a dummy value to P; since _m>=n_, the total time is in _O(m)_
- the table _P_ is iterated and for each element the tree is climbed to the root to compute the size of the subtree of each node stored in a table _S_; this operation requires _O(n*d)_ time, where _d_ is the height of the node, which is in average case much lower than m (d=m) is the worst case)
- next we construct the tree data _D_, consisting of the depth first order of the nodes, and the list _C_ of the coordinates of all nodes in the tree data, by id; to do this, we first add the root, then iterate again over _P_ , climb the tree from it, until the next not-yet-added ancestor is found, and the position where to add it is computed by the next free position in the subtree of its parent (which must have been already added, by definition, thus the next free position in its subtree is known). Although this requires climbing the tree, this step takes _O(n)_ time, since each node is added only once.

The last step was taking the most time in the first implementation, because the next free position was computed by iterating over the subtree sizes of the ancestors, thus the total time was _O(n*h)_, however it was reduced to _O(n)_ in the new implementation.

Now the slowest step is the second (computation of _S_). Since each node must be count in the subtree size of all its ancestors, there is no way to reduce the time from _O(n*h)_.

To parallelize this step, one divides the parents table into _t_ slices, and assign each to a different process (not thread, because of the GIL). Each process would then count the subtree sizes in the slice only. I implemented two versions. In the first one I used a shared table and a lock, but this was very bad performing and I didn't keep it. In the second one, each process makes its own S table and the results are then summed up. This version is now in the main branch and can be activated in the CLI, with an option ``--processes``.

The parallel version turns out to be slower than the serial one, for any number of processes, when constructing the NCBI tree. The overhead of the process starting, initializing the arrays and summing up the results is too high.

For example, on my system:
```
ggmac22:fastsubtrees gonnella$ time fastsubtrees tree nt.tree fastsubtrees/ntdumps/nodes.dmp --ncbi --force --processes 8
2022-10-20 18:39:43 INFO: Constructing temporary parents table...
2022-10-20 18:39:43 INFO: Reading data from file "fastsubtrees/ntdumps/nodes.dmp" ...
2449598it [00:02, 1013332.31it/s]
2022-10-20 18:39:45 INFO: Constructing subtree sizes table using 8 processes...
2022-10-20 18:40:01 INFO: Computing depth-first tree traversal order...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:01<00:00, 1603013.94it/s]
2022-10-20 18:40:03 INFO: Finalize index of nodes positions in depth-first traversal...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:00<00:00, 4386400.63it/s]
2022-10-20 18:40:03 SUCCESS: Tree data structure constructed
2022-10-20 18:40:03 INFO: Tree written to file "nt.tree"

real	0m21.018s
user	0m24.954s
sys	0m1.279s
ggmac22:fastsubtrees gonnella$ time fastsubtrees tree nt.tree fastsubtrees/ntdumps/nodes.dmp --ncbi --force --processes 4
2022-10-20 18:40:06 INFO: Constructing temporary parents table...
2022-10-20 18:40:06 INFO: Reading data from file "fastsubtrees/ntdumps/nodes.dmp" ...
2449598it [00:02, 1049673.20it/s]
2022-10-20 18:40:08 INFO: Constructing subtree sizes table using 4 processes...
2022-10-20 18:40:17 INFO: Computing depth-first tree traversal order...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:01<00:00, 1609707.83it/s]
2022-10-20 18:40:19 INFO: Finalize index of nodes positions in depth-first traversal...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:00<00:00, 4494199.61it/s]
2022-10-20 18:40:20 SUCCESS: Tree data structure constructed
2022-10-20 18:40:20 INFO: Tree written to file "nt.tree"

real	0m14.172s
user	0m17.283s
sys	0m0.571s
ggmac22:fastsubtrees gonnella$ time fastsubtrees tree nt.tree fastsubtrees/ntdumps/nodes.dmp --ncbi --force --processes 1
2022-10-20 18:40:25 INFO: Constructing temporary parents table...
2022-10-20 18:40:25 INFO: Reading data from file "fastsubtrees/ntdumps/nodes.dmp" ...
2449598it [00:02, 1045972.70it/s]
2022-10-20 18:40:27 INFO: Constructing subtree sizes table...
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:09<00:00, 330047.63it/s]
2022-10-20 18:40:37 INFO: Computing depth-first tree traversal order...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:01<00:00, 1634011.31it/s]
2022-10-20 18:40:38 INFO: Finalize index of nodes positions in depth-first traversal...
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████| 2987600/2987600 [00:00<00:00, 4426591.91it/s]
2022-10-20 18:40:39 SUCCESS: Tree data structure constructed
2022-10-20 18:40:39 INFO: Tree written to file "nt.tree"

real	0m14.258s
user	0m14.044s
sys	0m0.207s
```
