### 2022-10-20 d36a65c
fix example in README
