### 2022-10-20 ggonnella/fastsubtrees#15 @ggonnella
OK it looks like a a MacOS problem. The standard tools on MacOS need the option before the other arguments. I'll turn the order.
