### 2022-10-20 eef44fd
fix genome_attributes_viewer prepare_attributes
