+++ genomes_attributes_viewer/CHANGELOG.txt
@@ -1,3 +1,7 @@
+== 1.3 ==
+
+- first run preparation was fixed
+
 == 1.2 ==
 - update to changes in fastsubtrees
 
