+++ genomes_attributes_viewer/CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Giorgio"
   orcid: "https://orcid.org/0000-0003-3900-5397"
 title: "Genomes attributes viewer: an example application for fastsubtrees"
-version: 1.0.0
+version: 1.3
 date-released: 2022-04-22
 url: "htts://github.com/ggonnella/fastsubtrees/tree/main/genomes_attributes_viewer"
