### 2022-10-20 33f6842
turn around options order in sh. calls for MacOS compatibility
