### 2022-10-20 ggonnella/fastsubtrees#15 @ggonnella
33f6842 fixes this
