+++ ntdownload/ntdownload/__init__.py
@@ -1,6 +1,6 @@
 # ntdownload library
 
-__version__="1.7"
+__version__="1.7.1"
 
 from .downloader import Downloader
 from .names import yield_scientific_names_from_dump
