+++ ntdownload/CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Giorgio"
   orcid: "https://orcid.org/0000-0003-3900-5397"
 title: "NtDownload: easily updatable local NCBI taxonomy dump files copy"
-version: 1.0.0
-date-released: 2022-04-22
+version: 1.7.1
+date-released: 2022-10-21
 url: "htts://github.com/ggonnella/fastsubtrees/tree/main/ntdownload"
