### 2022-10-20 9a87a0a
remove duplicated command in docker examples
