### 2022-10-20 66d69c1
make sure that all examples in README.md run
