### 2022-10-20 d35315c
add tree construction algorithm notes to README
