### 2022-10-20 ggonnella/fastsubtrees#8 @ggonnella
I added most of this information to the README.md and a sentence describing the parallel version to the paper.
