### 2022-10-20 bc1ff5d
add a docstring to the Tree class
