### 2022-10-20 #4755 @ggonnella
@vinisalazar:
- I added a docstring to the ``Tree`` class (the only public class of the library). Public functions mentioned in the API manual already have a docstring and even type annotations.
- I fixed the problem with the tests on ntdownload, which was related to the MacOSX version of the Posix tools.
- I double checked and fixed the examples in README.md, which are now all runnable.
