### 2022-10-20 #4755 @ggonnella
@KonradHoeffner:
- I added a sentence to the paper briefly describing the parallel implementation of the tree construction, with a more detailed description in the library README
- I fixed the problems with genomes_attributes_viewer
