### 2022-10-21 35f2cbc
better wording in tree construction algorithm description
