### 2022-10-21 #4755 @KonradHoeffner
A few comments from my final read:

* *example are provided* -> *example**s***
* *no Python package currently allows*: are there packages for other programming languages?
* *with Apple M1 Pro CPU* -> *with **an** Apple M1 Pro CPU* or *with **the** Apple M1 Pro CPU* (I'm not an English native speaker but it sounds better to me with an article)
* *better scales* -> *scales better*
* If possible, align the numbers in the tables so that each digit with the same significance has the same horizontal position (i.e. right aligned for integers or numbers with the same numbers of decimal places).
* write out the meaning of *GC* on the first usage
