### 2022-10-21 ggonnella/fastsubtrees#8 @KonradHoeffner
Thank you for the detailed explanation!
