### 2022-10-21 93f9eeb
fixed minor issues listed by KonradHoeffner; removed repository cit; added versions and pip commands for all subpackages
