+++ docs/paper.md
@@ -57,7 +57,7 @@ in multiple subtrees, which are compared to each other.
 This allows for instance to identify uncommon values in some members of a taxon,
 which in turn can be used in a primer for new biological hypotheses.
 
-No Python package currently allows for a simple and efficient extraction of a
+No package currently allows for a simple and efficient extraction of a
 subtree, including data associated with the subtree nodes.
 While software exists for visualizing data associated with phylogenetic
 or taxonomic subtrees, e.g. AnnoTree [@Mendler:2019] and Treehouse
@@ -66,7 +66,7 @@ to integrate in other applications.
 
 # The fastsubtrees Python package
 
-Here we present the  _fastsubtrees_ Python package [@FastsubtreesRepo], which was developed
+Here we present the  _fastsubtrees_ Python package which was developed
 for storing tree information using a compact representation,
 suitable for fast extraction of any subtree.
 This package can be applied to any tree in which the nodes
@@ -75,7 +75,7 @@ tested mainly for the NCBI taxonomy tree, it is not limited to it.
 
 It can be installed using ``pip install fastsubtrees``.
 The package functionality can be used
-through either executable scripts or an API, which can be employed from other
+through either a command line script or an API, which can be employed from other
 Python programs. Manuals are provided for both kind of interfaces. A complete
 test suite based on _pytest_ is provided. Running the complete tests, comparative
 benchmarks and the example application (see below) requires a database
@@ -84,9 +84,10 @@ to simply run benchmarks, tests and the example application without this hassle.
 
 The first step when using _fastsubtrees_ is constructing its tree
 representation and storing it to file. Any source can be used as input (e.g. tabular files, or
-database tables) by defining a generation function (example are provided).
-A separate wrapper package _ntsubtree_, installable by ``pip``,
-is provided, which simplifies working with the NCBI taxonomy tree.
+database tables) by defining a generation function (examples are provided).
+A separate wrapper package _ntsubtree_, installable by ``pip install ntsubtree``,
+is provided (version 1.1; directory ``ntsubtree`` of the source code repository),
+which simplifies working with the NCBI taxonomy tree.
 Besides the NCBI taxonomy tree, other trees can be represented using
 _fastsubtrees_. For this, each node must be labeled by a unique non-negative integer ID.
 Furthermore, the IDs space
@@ -106,7 +107,7 @@ dynamic, i.e. it is possible to add and remove subtrees.
 
 # Subtree extraction benchmarks
 
-Benchmarks were performed on a MacBook Pro 2021 with Apple M1 Pro CPU and 32 Gb
+Benchmarks were performed on a MacBook Pro 2021 with an Apple M1 Pro CPU and 32 Gb
 RAM. Running times were measured as an average of 3 runs using GNU time version
 1.9 [@GNUtime]. Thereby, CPU time was computed as the sum of user and system
 time. For the tests Python version 3.10.2 was used. The tested version of
@@ -115,9 +116,9 @@ fastsubtrees was 2.0.
 The NCBI taxonomy tree used for the tests was downloaded on October 7, 2022
 from the NCBI FTP website [@NCBI:FTP].
 For downloading and keeping up-to-date a copy of the dump files using Python
-we developed the _ntdownload_ package (located in the directory ``ntdownload``
-of the source code repository, version 1.4)
-[@NTDOWNLOAD] installable using ``pip install ntdownload``.
+we developed the _ntdownload_ package (version 1.7; directory ``ntdownload``
+of the source code repository)
+installable using ``pip install ntdownload``.
[…]
