+++ docs/paper.bib
@@ -25,33 +25,6 @@
   url = {https://dash.plotly.com/}
 }
 
-@misc{NTMIRROR,
-  author = {Giorgio Gonnella},
-  title = {NtMirror Python package},
-  year = {2022},
-  publisher = {Github},
-  journal = {Github repository},
-  url = {https://github.com/ggonnella/fastsubtrees/tree/main/ntmirror}
-}
-
-@misc{NTDOWNLOAD,
-  author = {Giorgio Gonnella},
-  title = {NtDownload Python package},
-  year = {2022},
-  publisher = {Github},
-  journal = {Github repository},
-  url = {https://github.com/ggonnella/fastsubtrees/tree/main/ntdownload}
-}
-
-@misc{FastsubtreesRepo,
-  author = {Aman Modi and Giorgio Gonnella},
-  title = {Fastsubtrees Python package},
-  year = {2022},
-  publisher = {Github},
-  journal = {Github repository},
-  url = {https://github.com/ggonnella/fastsubtrees}
-}
-
 @article{Mendler:2019,
     author = {Mendler, Kerrin and Chen, Han and Parks, Donovan H and Lobb, Briallen and Hug, Laura A and Doxey, Andrew C},
     title = "{AnnoTree: visualization and exploration of a functionally annotated microbial tree of life}",
