### 2022-10-21 #4755 @ggonnella
> A few comments from my final read:
>
> * _example are provided_ -> _example**s**_

Fixed, thank you

> * _no Python package currently allows_: are there packages for other programming languages?

I removed "Python" from the sentence. No, I am not aware of any package, also for other programming languages.

> * _with Apple M1 Pro CPU_ -> _with **an** Apple M1 Pro CPU_ or _with **the** Apple M1 Pro CPU_ (I'm not an English native speaker but it sounds better to me with an article)

I added _an_. Thank you.

> * _better scales_ -> _scales better_

Changed. Thank you.

> * If possible, align the numbers in the tables so that each digit with the same significance has the same horizontal position (i.e. right aligned for integers or numbers with the same numbers of decimal places).

I don't think this is possible in Markdown.

> * write out the meaning of _GC_ on the first usage

I did it now. Thank you.

Additionally I removed the self citation to the URL of fastsubtrees Git repository and of ntdownload subdirectory. I don't think it's necessary, since the paper PDF already contains a link. Instead I added to the text the ``pip`` installation command, the current version and the subdirectory for each of the mentioned sub-packages.
