### 2022-10-21 8b16383
made fastsubtrees always italics in paper
