### 2022-10-21 #4755 @KonradHoeffner
It's a pity that the word count is limited for JOSS papers because the algorithm description and extended discussion of parallelism in https://github.com/ggonnella/fastsubtrees/issues/8 is really interesting and it would be really nice to have that in the paper itself.
