### 2022-10-21 #4755 @ggonnella
(I found another typo and unified the formatting of _fastsubtrees_ in the text to italics)
