+++ fastsubtrees/__init__.py
@@ -1,6 +1,6 @@
 # Fastsubtrees library
 
-__version__="2.0"
+__version__="2.1"
 
 # setup loguru by disabling it, as expected for libraries
 from loguru import logger
