+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Giorgio"
   orcid: "https://orcid.org/0000-0003-3900-5397"
 title: 'Fastsubtrees: simple and efficient subtrees extractions in Python with applications to NCBI taxonomy'
-version: 2.0
+version: 2.1
 date-released: 2022-10-12
 url: "htts://github.com/ggonnella/fastsubtrees/"
