### 2022-10-21 5b1ea1c
mention version 2.1 in the paper
(since parallel computation was mentioned, which was introduced in 2.1)
