### 2022-10-21 e6d08e1
formatting the list of required arrays
