### 2022-10-21 #4755 @ggonnella
> It's a pity that the word count is limited for JOSS papers because the algorithm description and extended discussion of parallelism in [ggonnella/fastsubtrees#8](https://github.com/ggonnella/fastsubtrees/issues/8) is really interesting and it would be really nice to have that in the paper itself.

I would also be glad if it would be possible to add it to the paper. Since both reviewer seemed to agree on this,  I polished that text, made it slightly more compact and added to the paper in a separate branch. This would be the resulting paper:

https://github.com/ggonnella/fastsubtrees/suites/8895052593/artifacts/406895074
