### 2022-10-21 34a877d
polish _compute_subtree_sizes_partial function code
