### 2022-10-21 c265894
try to make text more compact
