### 2022-10-27 #4755 @ggonnella
> @KonradHoeffner thanks for your initial review.
>
> The JOSS review process is slightly different from a traditional paper review in that you'll use this Gitlab issue and/or issues on the software repository to communicate with the authors as you work though the review. So far, you've checked boxes which have been sufficiently addressed, and you've posted here about missing features, which is exactly how the process starts. As the authors work to address the issues and as you confirm the issues have been addressed, you'll check off more and more checkboxes until all requirements have been met.
>
> Once both reviewers on the paper have checked off all of the boxes in their respective checklists, I'll look over the paper and run through some finalization steps.
>
> Hope that makes sense - feel free to ping me again if you have any questions or concerns.

@kellyrowland  As author I can't tell exactly, but from the comments of the two reviewers, see above, I imagine that all the boxes are now checked? How will this go on now? ... (It remains open the question about the additional paragraph, about parallelisation wrote after request from the reviewers, which could be gladly added, please see discussion above.)
