### 2022-11-10 #4755 @kellyrowland
Thanks all for the work here; apologies for the delay on my end.

@ggonnella @KonradHoeffner I've looked at the differences between the two branches of `paper.md` and I agree that the parallelization discussion really adds to the paper, so I would recommend including that (it's okay that it's a bit over the recommended word count).

If you could please ping me again here when the the changes are merged to `main`, I'll build the new PDF and then we can proceed with wrapping up the review.
