### 2022-11-11 #4755 @KonradHoeffner
Thank you @kellyrowland!
The final article proof looks great, I only found a few minor issues:

* 32 Gb RAM -> should probably be 32 GB (PC memory is usually stated in Gigabytes, not Gigabits)
* "The authors would like to thank Serena Lam (Department of Bioinformatics, University of
Göttingen) for language style suggestions and grammar corrections" -> add missing full stop to the end
* "Dash python" -> "Dash Python" (Python is a proper name)
