### 2022-11-11 d5bbcfe
paper minor corrections spotted by reviewer Konrad Hoeffner
