+++ docs/paper.bib
@@ -18,7 +18,7 @@
 
 @misc{Dash,
   author = {},
-  title = {Dash Python User Guide},
+  title = {Dash {P}ython User Guide},
   year = {2022},
   publisher = {plotly},
   journal = {Dash Project Website},
