### 2022-11-13 #4755 @Kevin-Mattheus-Moerman
@ggonnella I am the AEiC for this track and here to help process the final steps towards acceptance in JOSS. I have just proofread your paper, and reviewed the ZENODO archive. Please work on the below minor issues (Feel free to implement different wordings from what I suggest, also feel free to disagree with or skip the recommendations):

- [ ] Please edit the ZENODO archive manually so that the title and the author matches those of the paper.
- [ ] It looks like `..the extraction of distribution of ...` needs an `s` to make distributions plural i.e.  `..the extraction of distributions of ...`
- [ ] Recommendation: consider rephrasing `This allows for instance to identify uncommon values in some...` to `This, for instance, allows for the identification of uncommon values in some...`
- [ ] I would avoid starting a sentence with `i.e.`, i.e. instead of `I.e., although it was designed and tested...` consider something like `Note however that, although it was designed and tested...`, if I understood correctly what you were trying to say.
- [ ] I would add commas in `The genome size attribute file generation using the fastsubtrees attribute command required...` such that it reads `The genome size attribute file generation, using the fastsubtrees attribute command, required...`
- [ ] The numbered list after `The construction algorithm consists in the following 3 operations:` does not appear correctly in the PDF, perhaps you require additional spaces in the markdown at the end of those list sentences to force the newline in markdown?
- [ ] Check `in worst case h = m, in average it is much smaller`, it looks like it should be: `in the worst case h = m, however, in general it is much smaller`. If you do mean `average` then `on average` is better.
- [ ] Recommendation: I wonder if you can work on getting GitHub to recognize the license you are using. I noticed you have some custom text at the top, i.e.

> `All code of fastsubtrees is released under the following ISC license. It is functionally equivalent to a two-term BSD copyright with language removed that is made unnecessary by the Berne convention. See http://openbsd.org/policy.html for more information on copyrights.`
Perhaps if you stuck with this exact wording GitHub would pick it up: https://opensource.org/licenses/ISC. It probably does not matter if you use a plain text LICENSE file (without extension), but that could be another thing to try to see if GitHub then scans it properly.
Note this is just a comment, this is not a requirement for publication. I merely had to check your license and since it did not seem detected I had a closer look.

- [ ] Recommendation: I recommend adding a note/link on the license in your README as well. You may also consider adding a license badge (see also: https://gist.github.com/ulises-jeremias/ebcacbc50c64d4e04bc8b161e2fa44a9). Note that the text I suggested to remove could be added instead to your README where you mention the license.

Let me know when you're done processing the above.
