### 2022-11-13 c78736d
LICENSE: move initial note to README and rm file extension
