### 2022-11-13 5bf5eea
format all badges on a single line
