### 2022-11-13 #4755 @ggonnella
I processed all the points in the list

* [X] Please edit the ZENODO archive manually so that the title and the author matches those of the paper.
* [X]  It looks like `..the extraction of distribution of ...` needs an `s` to make distributions plural i.e.  `..the extraction of distributions of ...`
   * I changed it from `...the extraction of distribution of values...` to ``...the extraction of values distribution...``.
* [X]  Recommendation: consider rephrasing `This allows for instance to identify uncommon values in some...` to `This, for instance, allows for the identification of uncommon values in some...`
* [X]  I would avoid starting a sentence with `i.e.`, i.e. instead of `I.e., although it was designed and tested...` consider something like `Note however that, although it was designed and tested...`, if I understood correctly what you were trying to say.
  * I removed the `I.e.` from the beginning of the sentence and added a `thus` to the main clause.
* [X]  I would add commas in `The genome size attribute file generation using the fastsubtrees attribute command required...` such that it reads `The genome size attribute file generation, using the fastsubtrees attribute command, required...`
* [X]  The numbered list after `The construction algorithm consists in the following 3 operations:` does not appear correctly in the PDF, perhaps you require additional spaces in the markdown at the end of those list sentences to force the newline in markdown?
* [X]  Check `in worst case h = m, in average it is much smaller`, it looks like it should be: `in the worst case h = m, however, in general it is much smaller`. If you do mean `average` then `on average` is better.
* [X]  Recommendation: I wonder if you can work on getting GitHub to recognize the license you are using. I noticed you have some custom text at the top
  * I removed the text and put it into README and changed the filename removing the extension .txt
  * Github recognizes the license now
* [X]  Recommendation: I recommend adding a note/link on the license in your README as well. You may also consider adding a license badge (see also: https://gist.github.com/ulises-jeremias/ebcacbc50c64d4e04bc8b161e2fa44a9). Note that the text I suggested to remove could be added instead to your README where you mention the license.
  * I put the removed text from LICENCE to README and added some batches, including the mentioned one
