### 2022-11-13 36b309f
fix Makefile clean goal
